Rename the benchmark algorithms, include header and comments
