--- Algorithm.CSharp/Benchmarks/EmptyMinute400EquityAlgorithm.cs
@@ -1,7998 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using QuantConnect.Data;
-
-namespace QuantConnect.Algorithm.CSharp.Benchmarks
-{
-    public class EmptyMinute400EquityAlgorithm : QCAlgorithm
-    {
-        public override void Initialize()
-        {
-            SetStartDate(2015, 09, 28);
-            SetEndDate(2015, 11, 13);
-            SetBenchmark(dt => 1m);
-            foreach (var symbol in Symbols.Equity.All.Take(400))
-            {
-                AddSecurity(SecurityType.Equity, symbol);
-            }
-        }
-
-        public override void OnData(Slice slice)
-        {
-        }
-
-        // included here to make running this in automated fashion easier
-        private static class Symbols
-        {
-            public static class Equity
-            {
-                #region All
-
-                public static readonly HashSet<string> All = new HashSet<string>
-                {
-                    "SPY",
-                    "AAPL",
-                    "FB",
-                    "VXX",
-                    "VRX",
-                    "NFLX",
-                    "UVXY",
-                    "QQQ",
-                    "IWM",
-                    "BABA",
-                    "GILD",
-                    "XIV",
-                    "XOM",
-                    "CVX",
-                    "MSFT",
-                    "GE",
-                    "SLB",
-                    "JPM",
-                    "XLE",
-                    "DIS",
-                    "AMZN",
-                    "TWTR",
-                    "PFE",
-                    "C",
-                    "BAC",
-                    "ABBV",
-                    "JNJ",
-                    "HAL",
-                    "XLV",
-                    "INTC",
-                    "WFC",
-                    "V",
-                    "YHOO",
-                    "COP",
-                    "MYL",
-                    "AGN",
-                    "WMT",
-                    "KMI",
-                    "MRK",
-                    "TSLA",
-                    "GDX",
-                    "LLY",
-                    "FCX",
-                    "CAT",
-                    "CELG",
-                    "QCOM",
-                    "MCD",
-                    "CMCSA",
-                    "XOP",
-                    "CVS",
-                    "AMGN",
-                    "DOW",
-                    "AAL",
-                    "APC",
-                    "SUNE",
-                    "MU",
-                    "VLO",
-                    "SBUX",
-                    "WMB",
-                    "PG",
-                    "EOG",
-                    "DVN",
-                    "BMY",
-                    "APA",
-                    "UNH",
-                    "EEM",
-                    "IBM",
-                    "NKE",
-                    "T",
-                    "HD",
-                    "UNP",
-                    "DAL",
-                    "ENDP",
-                    "CSCO",
-                    "OXY",
-                    "MRO",
-                    "MDT",
-                    "TXN",
-                    "WLL",
-                    "ORCL",
-                    "GOOGL",
-                    "UAL",
-                    "WYNN",
-                    "MS",
-                    "HZNP",
-                    "BIIB",
-                    "VZ",
-                    "GM",
-                    "NBL",
-                    "TWX",
-                    "SWKS",
-                    "JD",
-                    "HCA",
-                    "AVGO",
-                    "YUM",
-                    "KO",
-                    "GOOG",
-                    "GS",
-                    "PEP",
-                    "AIG",
-                    "EMC",
-                    "BIDU",
-                    "CLR",
-                    "PYPL",
-                    "LVS",
-                    "SWN",
-                    "AXP",
-                    "ATVI",
-                    "RRC",
-                    "WBA",
-                    "MPC",
-                    "NXPI",
-                    "ETE",
-                    "NOV",
-                    "FOXA",
-                    "SNDK",
-                    "DIA",
-                    "UTX",
-                    "DD",
-                    "WDC",
-                    "AA",
-                    "M",
-                    "FXI",
-                    "RIG",
-                    "MA",
-                    "DUST",
-                    "TGT",
-                    "AET",
-                    "EBAY",
-                    "LUV",
-                    "EFA",
-                    "BRK.B",
-                    "BA",
-                    "MET",
-                    "LYB",
-                    "SVXY",
-                    "UWTI",
-                    "HON",
-                    "HPQ",
-                    "OAS",
-                    "ABT",
-                    "MO",
-                    "ESRX",
-                    "TEVA",
-                    "STX",
-                    "IBB",
-                    "F",
-                    "CBS",
-                    "TLT",
-                    "PM",
-                    "ESV",
-                    "NE",
-                    "PSX",
-                    "SCHW",
-                    "MON",
-                    "HES",
-                    "GPRO",
-                    "TVIX",
-                    "MNK",
-                    "NVDA",
-                    "NFX",
-                    "USO",
-                    "NUGT",
-                    "EWZ",
-                    "LOW",
-                    "UA",
-                    "TNA",
-                    "XLY",
-                    "MMM",
-                    "PXD",
-                    "VIAB",
-                    "MDLZ",
-                    "NEM",
-                    "USB",
-                    "MUR",
-                    "ETN",
-                    "FEYE",
-                    "PTEN",
-                    "OIH",
-                    "UPS",
-                    "CHK",
-                    "DHR",
-                    "RAI",
-                    "TQQQ",
-                    "CCL",
-                    "BRCM",
-                    "DG",
-                    "JBLU",
-                    "CRM",
-                    "ADBE",
-                    "COG",
-                    "PBR",
-                    "HP",
-                    "BHI",
-                    "BK",
-                    "TJX",
-                    "DE",
-                    "COF",
-                    "INCY",
-                    "DHI",
-                    "ABC",
-                    "XLI",
-                    "ZTS",
-                    "BP",
-                    "IYR",
-                    "PNC",
-                    "CNX",
-                    "XLF",
-                    "LRCX",
-                    "GG",
-                    "RDS.A",
-                    "WFM",
-                    "TSO",
-                    "ANTM",
-                    "KSS",
-                    "EA",
-                    "PRU",
-                    "RAD",
-                    "WFT",
-                    "XBI",
-                    "THC",
-                    "VWO",
-                    "CTSH",
-                    "ABX",
-                    "VMW",
-                    "CSX",
-                    "ACN",
-                    "EMR",
-                    "SE",
-                    "MJN",
-                    "SKX",
-                    "ACE",
-                    "P",
-                    "CMI",
-                    "CL",
-                    "CAH",
-                    "EXC",
-                    "DUK",
-                    "AMAT",
-                    "AEM",
-                    "FTI",
-                    "STT",
-                    "ILMN",
-                    "HOG",
-                    "KR",
-                    "EXPE",
-                    "VRTX",
-                    "IVV",
-                    "CAM",
-                    "GPS",
-                    "MCK",
-                    "ADSK",
-                    "CMCSK",
-                    "HTZ",
-                    "MGM",
-                    "DLTR",
-                    "STI",
-                    "CYH",
-                    "MOS",
-                    "CNQ",
-                    "GLW",
-                    "KEY",
-                    "KORS",
-                    "SIRI",
-                    "EPD",
-                    "SU",
-                    "DFS",
-                    "TMO",
-                    "TAP",
-                    "HST",
-                    "NBR",
-                    "EQT",
-                    "XLU",
-                    "BSX",
-                    "COST",
-                    "CTRP",
-                    "HFC",
-                    "VNQ",
-                    "TRV",
-                    "POT",
-                    "CERN",
-                    "LLTC",
-                    "DO",
-                    "ADI",
-                    "BAX",
-                    "AMT",
-                    "URI",
-                    "UCO",
-                    "ECA",
-                    "MAS",
-                    "ALL",
-                    "PCAR",
-                    "VIPS",
-                    "ATW",
-                    "SPXU",
-                    "LNKD",
-                    "X",
-                    "TSM",
-                    "SO",
-                    "BBT",
-                    "SYF",
-                    "VFC",
-                    "CXO",
-                    "IR",
-                    "PWR",
-                    "GLD",
-                    "LNG",
-                    "ETP",
-                    "JNPR",
-                    "MAT",
-                    "KLAC",
-                    "XLK",
-                    "TRIP",
-                    "AEP",
-                    "VTR",
-                    "ROST",
-                    "RDC",
-                    "CF",
-                    "FAS",
-                    "HCN",
-                    "AR",
-                    "SM",
-                    "WPX",
-                    "D",
-                    "HOT",
-                    "PRGO",
-                    "ALXN",
-                    "CNC",
-                    "VALE",
-                    "JCP",
-                    "GDXJ",
-                    "OKE",
-                    "ADM",
-                    "JOY",
-                    "TSN",
-                    "MAR",
-                    "KHC",
-                    "NSC",
-                    "CMA",
-                    "COH",
-                    "GMCR",
-                    "FL",
-                    "FITB",
-                    "BHP",
-                    "JWN",
-                    "DNR",
-                    "PBF",
-                    "XLNX",
-                    "PHM",
-                    "HIG",
-                    "PPG",
-                    "MBLY",
-                    "ITUB",
-                    "FDX",
-                    "IP",
-                    "EL",
-                    "LEN",
-                    "AFL",
-                    "HLT",
-                    "XRX",
-                    "NEE",
-                    "RCL",
-                    "NRG",
-                    "WTW",
-                    "QEP",
-                    "JAH",
-                    "SQQQ",
-                    "RIO",
-                    "CFG",
-                    "PAH",
-                    "XLB",
-                    "RSPP",
-                    "FOX",
-                    "LBTYK",
-                    "FAST",
-                    "XLP",
-                    "TZA",
-                    "KMB",
-                    "HCP",
-                    "JCI",
-                    "LMT",
-                    "SPG",
-                    "HBI",
-                    "BBY",
-                    "BMRN",
-                    "BEN",
-                    "GIS",
-                    "PCG",
-                    "WY",
-                    "RTN",
-                    "DISCA",
-                    "IVZ",
-                    "SLW",
-                    "TCK",
-                    "BWA",
-                    "NUE",
-                    "ADP",
-                    "CCI",
-                    "LNC",
-                    "FOLD",
-                    "DLPH",
-                    "BKD",
-                    "ISIS",
-                    "MTB",
-                    "S",
-                    "PPL",
-                    "PH",
-                    "ED",
-                    "EQR",
-                    "TMUS",
-                    "TEL",
-                    "RF",
-                    "CHTR",
-                    "GGP",
-                    "AMTD",
-                    "LULU",
-                    "XON",
-                    "SYY",
-                    "WNR",
-                    "BX",
-                    "NVAX",
-                    "FSLR",
-                    "BBBY",
-                    "MDVN",
-                    "PEG",
-                    "EIX",
-                    "STJ",
-                    "ITW",
-                    "CRZO",
-                    "O",
-                    "AKAM",
-                    "REGN",
-                    "GT",
-                    "VRSK",
-                    "RY",
-                    "BLUE",
-                    "NVS",
-                    "SYK",
-                    "HUN",
-                    "SSO",
-                    "KSU",
-                    "PE",
-                    "SDRL",
-                    "LBTYA",
-                    "RAX",
-                    "ECL",
-                    "SPN",
-                    "SCTY",
-                    "GD",
-                    "ETFC",
-                    "CTL",
-                    "PANW",
-                    "EXAS",
-                    "MCHP",
-                    "DYAX",
-                    "CAR",
-                    "PX",
-                    "TOL",
-                    "LVLT",
-                    "OPK",
-                    "SEE",
-                    "PCLN",
-                    "HOLX",
-                    "FIT",
-                    "CI",
-                    "XEC",
-                    "NRF",
-                    "PLD",
-                    "PAA",
-                    "FLS",
-                    "GPOR",
-                    "UPRO",
-                    "PAYX",
-                    "MXIM",
-                    "DOV",
-                    "ERX",
-                    "BXLT",
-                    "KMX",
-                    "AER",
-                    "K",
-                    "LB",
-                    "SPLS",
-                    "BUD",
-                    "GRUB",
-                    "NTAP",
-                    "KRE",
-                    "WWAV",
-                    "MMC",
-                    "PNR",
-                    "CTXS",
-                    "CY",
-                    "SMH",
-                    "TROW",
-                    "AMBA",
-                    "WEC",
-                    "GSK",
-                    "FTR",
-                    "CAG",
-                    "EVHC",
-                    "XEL",
-                    "AMP",
-                    "MTDR",
-                    "DKS",
-                    "SYMC",
-                    "QUNR",
-                    "BBD",
-                    "HRB",
-                    "FLR",
-                    "DXJ",
-                    "FE",
-                    "FMC",
-                    "TWC",
-                    "CEMP",
-                    "IDTI",
-                    "MGA",
-                    "RDS.B",
-                    "TXT",
-                    "TDC",
-                    "SPLK",
-                    "FIS",
-                    "FTNT",
-                    "SRE",
-                    "SRPT",
-                    "OLN",
-                    "YELP",
-                    "PGR",
-                    "CAB",
-                    "NTRS",
-                    "MPEL",
-                    "MDY",
-                    "VMC",
-                    "OMC",
-                    "SDS",
-                    "CPE",
-                    "TRN",
-                    "FNV",
-                    "DRI",
-                    "TERP",
-                    "AES",
-                    "NOC",
-                    "DYN",
-                    "QID",
-                    "CBG",
-                    "CMG",
-                    "QRVO",
-                    "ETR",
-                    "A",
-                    "EW",
-                    "LPI",
-                    "NLSN",
-                    "WM",
-                    "MHFI",
-                    "GNC",
-                    "ZION",
-                    "TS",
-                    "TSS",
-                    "INTU",
-                    "SHW",
-                    "HUM",
-                    "CFX",
-                    "HBAN",
-                    "EMN",
-                    "BDX",
-                    "CIT",
-                    "SLCA",
-                    "CX",
-                    "AEO",
-                    "CNP",
-                    "ALK",
-                    "WU",
-                    "NWL",
-                    "RL",
-                    "CIE",
-                    "MLM",
-                    "YNDX",
-                    "FFIV",
-                    "FANG",
-                    "SWFT",
-                    "NYCB",
-                    "STLD",
-                    "AON",
-                    "CRC",
-                    "DK",
-                    "EWJ",
-                    "TYC",
-                    "CBI",
-                    "ENB",
-                    "STZ",
-                    "ZBH",
-                    "ROK",
-                    "DGX",
-                    "CRUS",
-                    "CME",
-                    "SHPG",
-                    "NAVI",
-                    "CLDX",
-                    "SGY",
-                    "XL",
-                    "CPN",
-                    "DATA",
-                    "HAIN",
-                    "KIM",
-                    "AKRX",
-                    "DPS",
-                    "KATE",
-                    "CP",
-                    "ACAD",
-                    "BCEI",
-                    "MNST",
-                    "VOYA",
-                    "AVB",
-                    "RICE",
-                    "ABEV",
-                    "CMS",
-                    "ARMH",
-                    "LH",
-                    "VTI",
-                    "WYN",
-                    "ALB",
-                    "XRAY",
-                    "CHKP",
-                    "EGN",
-                    "OC",
-                    "SAVE",
-                    "HDS",
-                    "BLK",
-                    "OII",
-                    "RLYP",
-                    "RDUS",
-                    "JBL",
-                    "AEZS",
-                    "FISV",
-                    "AAP",
-                    "CNI",
-                    "DWTI",
-                    "SNI",
-                    "NBIX",
-                    "TIF",
-                    "WHR",
-                    "RHT",
-                    "ON",
-                    "FOSL",
-                    "QVCA",
-                    "WPZ",
-                    "ANF",
-                    "MSI",
-                    "HEDJ",
-                    "UHS",
-                    "SD",
-                    "AMX",
-                    "IPG",
-                    "MCO",
-                    "APD",
-                    "QLIK",
-                    "ASNA",
-                    "CPHD",
-                    "CA",
-                    "WETF",
-                    "KBH",
-                    "EAT",
-                    "VOD",
-                    "WRK",
-                    "HSY",
-                    "CVE",
-                    "CB",
-                    "DTE",
-                    "ALLY",
-                    "FBHS",
-                    "XRT",
-                    "DSW",
-                    "CCE",
-                    "XPO",
-                    "NLY",
-                    "RSX",
-                    "HYG",
-                    "DISH",
-                    "CPGX",
-                    "BBL",
-                    "INFN",
-                    "MRVL",
-                    "ES",
-                    "FCAU",
-                    "NVO",
-                    "AEE",
-                    "ALNY",
-                    "WDAY",
-                    "EXPD",
-                    "DDD",
-                    "MRD",
-                    "LABU",
-                    "GOLD",
-                    "PBCT",
-                    "VOO",
-                    "CPB",
-                    "CIEN",
-                    "SBAC",
-                    "ATI",
-                    "PFG",
-                    "GME",
-                    "CSIQ",
-                    "ERY",
-                    "XHB",
-                    "CHRW",
-                    "GWW",
-                    "DVA",
-                    "PDCE",
-                    "MHK",
-                    "TOT",
-                    "DISCK",
-                    "CLX",
-                    "HAS",
-                    "GRPN",
-                    "AME",
-                    "AU",
-                    "VAR",
-                    "TSCO",
-                    "EWW",
-                    "VNTV",
-                    "GPN",
-                    "PAGP",
-                    "NOW",
-                    "SWK",
-                    "OIS",
-                    "SRCL",
-                    "ANAC",
-                    "MTG",
-                    "DEPO",
-                    "ST",
-                    "WAB",
-                    "BURL",
-                    "DXCM",
-                    "MD",
-                    "DNKN",
-                    "CVC",
-                    "ALKS",
-                    "PCL",
-                    "MTW",
-                    "NWSA",
-                    "URBN",
-                    "MDCO",
-                    "RGLD",
-                    "AWAY",
-                    "RES",
-                    "UTHR",
-                    "EROS",
-                    "AMJ",
-                    "USG",
-                    "TER",
-                    "LCI",
-                    "EQIX",
-                    "JEC",
-                    "NUS",
-                    "PVH",
-                    "UDR",
-                    "BXP",
-                    "VNO",
-                    "NDAQ",
-                    "IRM",
-                    "UNT",
-                    "UPL",
-                    "AMLP",
-                    "SAP",
-                    "FTK",
-                    "PII",
-                    "AGCO",
-                    "AIV",
-                    "LC",
-                    "LHO",
-                    "AZN",
-                    "APH",
-                    "ADT",
-                    "LM",
-                    "DDR",
-                    "DLR",
-                    "SGEN",
-                    "FET",
-                    "ICE",
-                    "NI",
-                    "PTCT",
-                    "TRGP",
-                    "XYL",
-                    "PSA",
-                    "BEAV",
-                    "EPE",
-                    "INFY",
-                    "PBR.A",
-                    "STO",
-                    "IWD",
-                    "PCRX",
-                    "GNW",
-                    "SJM",
-                    "SCG",
-                    "TASR",
-                    "RSG",
-                    "OSK",
-                    "NCLH",
-                    "IACI",
-                    "TWOU",
-                    "HALO",
-                    "DAN",
-                    "FNFG",
-                    "CE",
-                    "GLNG",
-                    "HNT",
-                    "HRL",
-                    "L",
-                    "AMAG",
-                    "HK",
-                    "R",
-                    "ORLY",
-                    "YOKU",
-                    "TMH",
-                    "PNW",
-                    "SFUN",
-                    "CREE",
-                    "SBH",
-                    "MRC",
-                    "EXP",
-                    "BRFS",
-                    "MT",
-                    "AMSG",
-                    "IBN",
-                    "ZIOP",
-                    "EXR",
-                    "MEOH",
-                    "MOH",
-                    "SPR",
-                    "ACHN",
-                    "TGNA",
-                    "VGK",
-                    "SPXS",
-                    "IPXL",
-                    "AVP",
-                    "Q",
-                    "PACB",
-                    "UN",
-                    "KKR",
-                    "VEA",
-                    "NAV",
-                    "AXLL",
-                    "BLL",
-                    "GRMN",
-                    "ITB",
-                    "BERY",
-                    "KN",
-                    "TD",
-                    "SYRG",
-                    "IWF",
-                    "PDCO",
-                    "HLF",
-                    "TRP",
-                    "TV",
-                    "VIXY",
-                    "AHS",
-                    "SFM",
-                    "BID",
-                    "HRTX",
-                    "TDW",
-                    "DGAZ",
-                    "MW",
-                    "PMCS",
-                    "CDNS",
-                    "UNM",
-                    "RLGY",
-                    "ATML",
-                    "JNS",
-                    "OHI",
-                    "PAY",
-                    "BBRY",
-                    "JAZZ",
-                    "JBHT",
-                    "MMP",
-                    "ASML",
-                    "AGO",
-                    "RHI",
-                    "WR",
-                    "EWY",
-                    "ESPR",
-                    "COTY",
-                    "BG",
-                    "NOK",
-                    "MAC",
-                    "SPXL",
-                    "SPWR",
-                    "INSY",
-                    "HA",
-                    "CUBE",
-                    "BLMN",
-                    "SLM",
-                    "CALM",
-                    "CSC",
-                    "TTWO",
-                    "OI",
-                    "QLD",
-                    "LPX",
-                    "ARIA",
-                    "GPC",
-                    "XME",
-                    "VRSN",
-                    "EZU",
-                    "CAA",
-                    "MWE",
-                    "UGAZ",
-                    "GPK",
-                    "MTZ",
-                    "AGNC",
-                    "SVU",
-                    "AGIO",
-                    "AXTA",
-                    "LEA",
-                    "DB",
-                    "ABMD",
-                    "STE",
-                    "MDRX",
-                    "DRE",
-                    "SCCO",
-                    "DPLO",
-                    "DECK",
-                    "BBG",
-                    "ACHC",
-                    "TRMB",
-                    "QSR",
-                    "TEX",
-                    "SSYS",
-                    "LL",
-                    "LMCA",
-                    "EXEL",
-                    "LABD",
-                    "QIHU",
-                    "ALJ",
-                    "SABR",
-                    "AOS",
-                    "NKTR",
-                    "QGEN",
-                    "AVT",
-                    "EWBC",
-                    "COMM",
-                    "WSM",
-                    "SN",
-                    "FNF",
-                    "RJF",
-                    "BRCD",
-                    "ACM",
-                    "AUY",
-                    "SC",
-                    "BOFI",
-                    "KGC",
-                    "LUK",
-                    "AN",
-                    "ZNGA",
-                    "BAM",
-                    "ICON",
-                    "HRS",
-                    "PAAS",
-                    "FHN",
-                    "LEG",
-                    "RDN",
-                    "NTES",
-                    "MNKD",
-                    "WLK",
-                    "HAR",
-                    "PKG",
-                    "NSAM",
-                    "ALTR",
-                    "ZFGN",
-                    "AGU",
-                    "MKC",
-                    "CENX",
-                    "FLO",
-                    "KAR",
-                    "VIV",
-                    "ERJ",
-                    "BIB",
-                    "LGF",
-                    "NNN",
-                    "ITC",
-                    "IMGN",
-                    "CLB",
-                    "FLEX",
-                    "ODP",
-                    "TWM",
-                    "CDW",
-                    "BWLD",
-                    "TPX",
-                    "WEN",
-                    "EXPR",
-                    "MDR",
-                    "SXL",
-                    "PPC",
-                    "AMCX",
-                    "KRC",
-                    "JUNO",
-                    "KMT",
-                    "AVY",
-                    "MDC",
-                    "SIG",
-                    "W",
-                    "RLJ",
-                    "BNS",
-                    "AKS",
-                    "NCR",
-                    "BRS",
-                    "ALLE",
-                    "TTM",
-                    "BCR",
-                    "HSBC",
-                    "MACK",
-                    "STNG",
-                    "LLL",
-                    "DF",
-                    "KND",
-                    "IDXX",
-                    "SLG",
-                    "APOL",
-                    "DNOW",
-                    "CMC",
-                    "CC",
-                    "SNV",
-                    "SONC",
-                    "LKQ",
-                    "TMK",
-                    "MAN",
-                    "FSL",
-                    "IM",
-                    "IWN",
-                    "MYGN",
-                    "YPF",
-                    "TPH",
-                    "BRX",
-                    "COL",
-                    "IRWD",
-                    "IEMG",
-                    "FRC",
-                    "ARMK",
-                    "CLVS",
-                    "Z",
-                    "CSL",
-                    "SNY",
-                    "WUBA",
-                    "HOS",
-                    "WAL",
-                    "NOG",
-                    "IJH",
-                    "CRI",
-                    "NAT",
-                    "AJG",
-                    "BC",
-                    "SHO",
-                    "ICPT",
-                    "MSCI",
-                    "GPRE",
-                    "PCP",
-                    "EFX",
-                    "BMR",
-                    "BYD",
-                    "HIW",
-                    "CCK",
-                    "KNX",
-                    "CYBR",
-                    "SRTY",
-                    "CAVM",
-                    "ULTA",
-                    "EGO",
-                    "SERV",
-                    "POST",
-                    "KITE",
-                    "WWW",
-                    "CTAS",
-                    "DCT",
-                    "ALSN",
-                    "ACOR",
-                    "BUFF",
-                    "OGE",
-                    "FCS",
-                    "HLS",
-                    "LQ",
-                    "ALGN",
-                    "UMPQ",
-                    "AWK",
-                    "BF.B",
-                    "WBS",
-                    "PACW",
-                    "TKR",
-                    "JNK",
-                    "BMO",
-                    "ARRS",
-                    "JNUG",
-                    "PEB",
-                    "BCS",
-                    "SBGI",
-                    "SCI",
-                    "GNRC",
-                    "CCJ",
-                    "IMS",
-                    "SLV",
-                    "NRZ",
-                    "SYNA",
-                    "AAN",
-                    "CHD",
-                    "ODFL",
-                    "FAZ",
-                    "LAZ",
-                    "PDS",
-                    "EV",
-                    "CAKE",
-                    "FBR",
-                    "PDLI",
-                    "CHS",
-                    "VER",
-                    "LXRX",
-                    "TRI",
-                    "ROP",
-                    "RS",
-                    "CJES",
-                    "KBR",
-                    "VEEV",
-                    "VC",
-                    "DRH",
-                    "DPZ",
-                    "VALE.P",
-                    "CINF",
-                    "CSTM",
-                    "SNH",
-                    "IWO",
-                    "HLX",
-                    "AIZ",
-                    "PKI",
-                    "TSU",
-                    "EEP",
-                    "KOS",
-                    "TNK",
-                    "ZSPH",
-                    "GXP",
-                    "DEI",
-                    "INFI",
-                    "HTWR",
-                    "TBT",
-                    "OCN",
-                    "MDXG",
-                    "NGLS",
-                    "AMG",
-                    "CPT",
-                    "CPG",
-                    "SIX",
-                    "LNT",
-                    "ALDR",
-                    "BCE",
-                    "AEGR",
-                    "IWB",
-                    "GWR",
-                    "FLT",
-                    "PF",
-                    "ALR",
-                    "OKS",
-                    "LXK",
-                    "POR",
-                    "CLF",
-                    "FINL",
-                    "STWD",
-                    "INVN",
-                    "ITCI",
-                    "SCMP",
-                    "MFC",
-                    "TTPH",
-                    "SEAS",
-                    "BIG",
-                    "OZRK",
-                    "RTRX",
-                    "ARG",
-                    "COO",
-                    "STR",
-                    "GNTX",
-                    "CEQP",
-                    "BTE",
-                    "SRC",
-                    "SUPN",
-                    "CPA",
-                    "WDR",
-                    "BCRX",
-                    "ITT",
-                    "ATHM",
-                    "ESS",
-                    "HSIC",
-                    "ATO",
-                    "TAHO",
-                    "CFR",
-                    "TEN",
-                    "RH",
-                    "OREX",
-                    "SWC",
-                    "ADXS",
-                    "RCII",
-                    "EWT",
-                    "LPLA",
-                    "AWI",
-                    "SGYP",
-                    "AL",
-                    "WRI",
-                    "BAS",
-                    "PBYI",
-                    "ADPT",
-                    "CNO",
-                    "SEIC",
-                    "BLDR",
-                    "CBOE",
-                    "ASH",
-                    "PTC",
-                    "JMEI",
-                    "HPE",
-                    "GFI",
-                    "ATU",
-                    "ISRG",
-                    "KS",
-                    "AXL",
-                    "MBI",
-                    "PBI",
-                    "XOMA",
-                    "HL",
-                    "JDST",
-                    "BAH",
-                    "NLNK",
-                    "PAYC",
-                    "NUAN",
-                    "PNRA",
-                    "XNPT",
-                    "MHR",
-                    "LMCK",
-                    "SIVB",
-                    "BKU",
-                    "ACC",
-                    "SINA",
-                    "SCSS",
-                    "ARW",
-                    "WAT",
-                    "GES",
-                    "KING",
-                    "ACWI",
-                    "MFRM",
-                    "DY",
-                    "SWI",
-                    "ALV",
-                    "RMD",
-                    "FRT",
-                    "MNTA",
-                    "MEG",
-                    "TCO",
-                    "HR",
-                    "AZO",
-                    "SCO",
-                    "POM",
-                    "UL",
-                    "IMPV",
-                    "BTU",
-                    "BDSI",
-                    "FLIR",
-                    "GBX",
-                    "CS",
-                    "STRZA",
-                    "TSL",
-                    "DOX",
-                    "CBSH",
-                    "PEGI",
-                    "KEX",
-                    "SGMS",
-                    "LTRPA",
-                    "HCBK",
-                    "TXRH",
-                    "NBG",
-                    "ADS",
-                    "THRX",
-                    "FIVE",
-                    "MTOR",
-                    "FDC",
-                    "SSRI",
-                    "ELLI",
-                    "PIR",
-                    "ANET",
-                    "ATHN",
-                    "ARE",
-                    "SHOO",
-                    "HXL",
-                    "KERX",
-                    "MSGN",
-                    "IEX",
-                    "TFM",
-                    "CST",
-                    "BVN",
-                    "UNFI",
-                    "PRXL",
-                    "ORI",
-                    "RMBS",
-                    "LYV",
-                    "TCB",
-                    "RJET",
-                    "PFPT",
-                    "TE",
-                    "FMER",
-                    "WCG",
-                    "GCI",
-                    "CBL",
-                    "RARE",
-                    "HPT",
-                    "BPOP",
-                    "SNPS",
-                    "TUP",
-                    "SGMO",
-                    "AMD",
-                    "TLN",
-                    "GRA",
-                    "ISBC",
-                    "LPNT",
-                    "CLI",
-                    "SNA",
-                    "NYLD",
-                    "CXW",
-                    "SPXC",
-                    "RPTP",
-                    "LII",
-                    "AYI",
-                    "WCC",
-                    "FCE.A",
-                    "TSRO",
-                    "LQD",
-                    "MANH",
-                    "UTIW",
-                    "CHL",
-                    "SKT",
-                    "PGNX",
-                    "CRL",
-                    "GOGO",
-                    "HDB",
-                    "BDN",
-                    "OMI",
-                    "ABY",
-                    "MSCC",
-                    "CROX",
-                    "UGI",
-                    "WOOF",
-                    "INSM",
-                    "DRQ",
-                    "CNK",
-                    "KBE",
-                    "FPRX",
-                    "LOCK",
-                    "CRR",
-                    "GIL",
-                    "ARRY",
-                    "UNG",
-                    "RKUS",
-                    "EDU",
-                    "ZAYO",
-                    "LPT",
-                    "UGP",
-                    "BMS",
-                    "RRD",
-                    "HTA",
-                    "WERN",
-                    "ACAS",
-                    "IMAX",
-                    "JACK",
-                    "WSH",
-                    "ZBRA",
-                    "NMBL",
-                    "MPW",
-                    "ROVI",
-                    "EPI",
-                    "NUVA",
-                    "SPPI",
-                    "PHG",
-                    "RPM",
-                    "SUM",
-                    "H",
-                    "MAA",
-                    "ABB",
-                    "SBNY",
-                    "SEMG",
-                    "PTLA",
-                    "BR",
-                    "PVTB",
-                    "TROX",
-                    "VNDA",
-                    "SEM",
-                    "FMX",
-                    "BPL",
-                    "ASB",
-                    "WNC",
-                    "BITA",
-                    "NSM",
-                    "GIII",
-                    "ICLR",
-                    "SXC",
-                    "IVW",
-                    "RGEN",
-                    "INT",
-                    "TWO",
-                    "IBKR",
-                    "CBD",
-                    "EXH",
-                    "OFC",
-                    "NGD",
-                    "MELI",
-                    "RACE",
-                    "CSOD",
-                    "IAG",
-                    "EWC",
-                    "NSR",
-                    "AFSI",
-                    "MSM",
-                    "HMHC",
-                    "HMY",
-                    "EQC",
-                    "IFF",
-                    "RYN",
-                    "ZINC",
-                    "DWRE",
-                    "ERF",
-                    "CUDA",
-                    "DFT",
-                    "FAF",
-                    "EPC",
-                    "LAMR",
-                    "SSTK",
-                    "DDS",
-                    "SF",
-                    "CSAL",
-                    "PENN",
-                    "VAL",
-                    "CRS",
-                    "TVPT",
-                    "MIC",
-                    "CLNY",
-                    "EWG",
-                    "VR",
-                    "DAR",
-                    "REG",
-                    "EQY",
-                    "TK",
-                    "SEDG",
-                    "CATY",
-                    "CTLT",
-                    "NWS",
-                    "MLNX",
-                    "MDU",
-                    "TDG",
-                    "CHMT",
-                    "EPZM",
-                    "RCI",
-                    "DV",
-                    "ONCE",
-                    "IMO",
-                    "HPP",
-                    "FR",
-                    "DGI",
-                    "UBS",
-                    "WMGI",
-                    "HYH",
-                    "DRII",
-                    "WRB",
-                    "CVI",
-                    "PRAA",
-                    "YRCW",
-                    "ZG",
-                    "LECO",
-                    "ENV",
-                    "AERI",
-                    "POL",
-                    "CONN",
-                    "WIN",
-                    "TIME",
-                    "ZLTQ",
-                    "CBM",
-                    "JKS",
-                    "GLOG",
-                    "VTV",
-                    "FULT",
-                    "TRCO",
-                    "GTLS",
-                    "WTR",
-                    "VA",
-                    "VLY",
-                    "DVAX",
-                    "PODD",
-                    "BXS",
-                    "INO",
-                    "ARNA",
-                    "EDZ",
-                    "MMS",
-                    "MTH",
-                    "RNG",
-                    "CMRX",
-                    "SMCI",
-                    "PNM",
-                    "MXL",
-                    "GGB",
-                    "BRO",
-                    "GTN",
-                    "ERIC",
-                    "HW",
-                    "TGI",
-                    "PLCM",
-                    "IJR",
-                    "DHT",
-                    "PVA",
-                    "SCOR",
-                    "IPGP",
-                    "THS",
-                    "GAS",
-                    "SANM",
-                    "VSH",
-                    "CVA",
-                    "DEO",
-                    "NFG",
-                    "ELS",
-                    "BKS",
-                    "NTCT",
-                    "PZZA",
-                    "GERN",
-                    "ANSS",
-                    "WBC",
-                    "DVY",
-                    "AAXJ",
-                    "MIK",
-                    "UFS",
-                    "WAFD",
-                    "BIS",
-                    "ELGX",
-                    "APLE",
-                    "PLAY",
-                    "SIRO",
-                    "EQM",
-                    "GLPI",
-                    "FNB",
-                    "INGR",
-                    "CGNX",
-                    "CSTE",
-                    "ERII",
-                    "NGG",
-                    "RGC",
-                    "CRTO",
-                    "EPAM",
-                    "FII",
-                    "UVE",
-                    "TNET",
-                    "OUTR",
-                    "SOXX",
-                    "RDY",
-                    "EIGI",
-                    "CDE",
-                    "HBHC",
-                    "CLH",
-                    "PNY",
-                    "CIM",
-                    "AF",
-                    "CLNE",
-                    "BAP",
-                    "ING",
-                    "BAK",
-                    "SIAL",
-                    "FEZ",
-                    "DCI",
-                    "IEF",
-                    "GEL",
-                    "DWA",
-                    "N",
-                    "EWH",
-                    "LSCC",
-                    "SBS",
-                    "HAWK",
-                    "CSII",
-                    "SPNC",
-                    "BEE",
-                    "LVNTA",
-                    "PPS",
-                    "OUT",
-                    "CTB",
-                    "LDOS",
-                    "MBT",
-                    "LDRH",
-                    "STRP",
-                    "SBGL",
-                    "CIG",
-                    "FCH",
-                    "AZPN",
-                    "HMSY",
-                    "RPAI",
-                    "LSTR",
-                    "BWXT",
-                    "IPI",
-                    "OEF",
-                    "ENR",
-                    "LEAF",
-                    "OLED",
-                    "HSC",
-                    "OPHT",
-                    "ZEN",
-                    "PETX",
-                    "MIDD",
-                    "AGEN",
-                    "LAD",
-                    "TREX",
-                    "ATRA",
-                    "ACGL",
-                    "MDSO",
-                    "OVAS",
-                    "VEU",
-                    "ARO",
-                    "IPHI",
-                    "MWW",
-                    "PBA",
-                    "SHAK",
-                    "FNSR",
-                    "TGTX",
-                    "SNE",
-                    "NVR",
-                    "VRNT",
-                    "HRC",
-                    "SSNC",
-                    "NXST",
-                    "SLH",
-                    "ENS",
-                    "FLOW",
-                    "TMST",
-                    "EDR",
-                    "DMND",
-                    "CBRL",
-                    "RWM",
-                    "CLGX",
-                    "FRAN",
-                    "PDM",
-                    "AWH",
-                    "KEG",
-                    "LGND",
-                    "TCBI",
-                    "SLF",
-                    "ESRT",
-                    "YY",
-                    "PB",
-                    "XENT",
-                    "IVR",
-                    "PLCE",
-                    "RBA",
-                    "NTGR",
-                    "QLGC",
-                    "GMED",
-                    "LXP",
-                    "JLL",
-                    "SDOW",
-                    "ARCC",
-                    "NTRI",
-                    "ABG",
-                    "BRKR",
-                    "EVR",
-                    "WPG",
-                    "FCEL",
-                    "CVLT",
-                    "FI",
-                    "HIMX",
-                    "PHH",
-                    "CUZ",
-                    "VIAV",
-                    "FLDM",
-                    "SWHC",
-                    "TRUE",
-                    "HIBB",
-                    "LXU",
-                    "MFA",
-                    "PGRE",
-                    "ANIP",
-                    "NYLD.A",
-                    "AVXL",
-                    "MMSI",
-                    "WES",
-                    "CBT",
-                    "PWE",
-                    "TDS",
-                    "TECD",
-                    "CONE",
-                    "GGG",
-                    "DPM",
-                    "NVRO",
-                    "GOV",
-                    "NR",
-                    "ENLK",
-                    "RE",
-                    "EDC",
-                    "VG",
-                    "HII",
-                    "SUI",
-                    "TXMD",
-                    "PNK",
-                    "EGRX",
-                    "SAGE",
-                    "CM",
-                    "CXRX",
-                    "AMH",
-                    "SBRA",
-                    "AGI",
-                    "PSTG",
-                    "TREE",
-                    "FLTX",
-                    "DDM",
-                    "OA",
-                    "WPC",
-                    "KEYS",
-                    "CSG",
-                    "RVNC",
-                    "LGIH",
-                    "RXN",
-                    "JOE",
-                    "RGLS",
-                    "WCN",
-                    "BWP",
-                    "VAC",
-                    "ZUMZ",
-                    "ENTA",
-                    "BZH",
-                    "VVC",
-                    "EWA",
-                    "CDK",
-                    "LOCO",
-                    "BXMT",
-                    "HDP",
-                    "SH",
-                    "AXS",
-                    "PRTA",
-                    "AG",
-                    "MLHR",
-                    "SAN",
-                    "RYAAY",
-                    "CASY",
-                    "AEL",
-                    "HTS",
-                    "BLOX",
-                    "DBEF",
-                    "EC",
-                    "ELNK",
-                    "GDDY",
-                    "SEP",
-                    "GLUU",
-                    "ETSY",
-                    "MENT",
-                    "UBSI",
-                    "CMCM",
-                    "CLD",
-                    "TUMI",
-                    "DOC",
-                    "STAG",
-                    "NGL",
-                    "GRFS",
-                    "NYRT",
-                    "ONB",
-                    "SSL",
-                    "PEIX",
-                    "BSMX",
-                    "CCOI",
-                    "CPRT",
-                    "GLBL",
-                    "SFLY",
-                    "MGNX",
-                    "STKL",
-                    "CG",
-                    "MPLX",
-                    "RYAM",
-                    "CHSP",
-                    "SSI",
-                    "GEO",
-                    "ANGI",
-                    "BLD",
-                    "RSP",
-                    "TRMK",
-                    "IDA",
-                    "HSNI",
-                    "TLLP",
-                    "DBD",
-                    "SNCR",
-                    "ASHR",
-                    "TEF",
-                    "APO",
-                    "MKTO",
-                    "SON",
-                    "STL",
-                    "TRVN",
-                    "MDAS",
-                    "GIMO",
-                    "RNR",
-                    "CLC",
-                    "SKYW",
-                    "ISIL",
-                    "ULTI",
-                    "MUSA",
-                    "EVER",
-                    "GWRE",
-                    "THO",
-                    "AIRM",
-                    "WOR",
-                    "GSAT",
-                    "TYL",
-                    "CMLS",
-                    "EEFT",
-                    "G",
-                    "CCP",
-                    "GMT",
-                    "BOX",
-                    "FDS",
-                    "TSRA",
-                    "SAFM",
-                    "IT",
-                    "XRS",
-                    "SSS",
-                    "NS",
-                    "TAL",
-                    "FMBI",
-                    "RUSL",
-                    "YINN",
-                    "JKHY",
-                    "AMKR",
-                    "HAE",
-                    "SNTA",
-                    "LNCE",
-                    "CUK",
-                    "RRGB",
-                    "EXAM",
-                    "VECO",
-                    "MSG",
-                    "KRG",
-                    "BGCP",
-                    "ZGNX",
-                    "SALE",
-                    "VIP",
-                    "ACI",
-                    "FIX",
-                    "HT",
-                    "UBNT",
-                    "WEX",
-                    "ABCO",
-                    "TILE",
-                    "GBSN",
-                    "MBFI",
-                    "GBCI",
-                    "FNGN",
-                    "NYT",
-                    "WWD",
-                    "EPR",
-                    "QLYS",
-                    "IHS",
-                    "SHLD",
-                    "EXXI",
-                    "CVBF",
-                    "MGLN",
-                    "EMES",
-                    "TTI",
-                    "MTSI",
-                    "ATR",
-                    "TMHC",
-                    "KLXI",
-                    "AMED",
-                    "PFF",
-                    "IEFA",
-                    "ESNT",
-                    "XXIA",
-                    "FGEN",
-                    "PINC",
-                    "DXD",
-                    "FOE",
-                    "TDOC",
-                    "CR",
-                    "VSI",
-                    "EURN",
-                    "TECH",
-                    "IGT",
-                    "TRQ",
-                    "JCOM",
-                    "MYCC",
-                    "ECR",
-                    "PES",
-                    "AGG",
-                    "ASPS",
-                    "ROIC",
-                    "LOGM",
-                    "SMTC",
-                    "ARWR",
-                    "CNL",
-                    "XCO",
-                    "SCHN",
-                    "NTI",
-                    "ARAY",
-                    "CYN",
-                    "DDC",
-                    "CTCT",
-                    "BKE",
-                    "CVG",
-                    "SNX",
-                    "BCC",
-                    "UE",
-                    "TFX",
-                    "STMP",
-                    "INCR",
-                    "NWE",
-                    "WLH",
-                    "WWE",
-                    "WTFC",
-                    "OFG",
-                    "VWR",
-                    "HPY",
-                    "HUB.B",
-                    "PBY",
-                    "ANTH",
-                    "VVUS",
-                    "CXP",
-                    "SCS",
-                    "IYE",
-                    "PGN",
-                    "VGR",
-                    "WBMD",
-                    "BSBR",
-                    "SJR",
-                    "SHLX",
-                    "CNHI",
-                    "FRGI",
-                    "AUO",
-                    "PAG",
-                    "EME",
-                    "VDSI",
-                    "KRA",
-                    "NXTM",
-                    "MKTX",
-                    "PKX",
-                    "CYS",
-                    "INXN",
-                    "CCRN",
-                    "RGA",
-                    "CALD",
-                    "WSO",
-                    "MASI",
-                    "MTD",
-                    "CYTK",
-                    "SGNT",
-                    "IGE",
-                    "RBC",
-                    "GWB",
-                    "WLB",
-                    "BJRI",
-                    "CZR",
-                    "EBS",
-                    "SFL",
-                    "IQNT",
-                    "STOR",
-                    "HEES",
-                    "POOL",
-                    "WGL",
-                    "AHT",
-                    "BGG",
-                    "PLUG",
-                    "VMI",
-                    "DNB",
-                    "SAAS",
-                    "TGH",
-                    "COLM",
-                    "BBCN",
-                    "ENH",
-                    "INDA",
-                    "AAVL",
-                    "ACIW",
-                    "PMT",
-                    "CIB",
-                    "TW",
-                    "OIBR",
-                    "POZN",
-                    "KALU",
-                    "GPI",
-                    "BBVA",
-                    "FDN",
-                    "GPT",
-                    "WST",
-                    "BOOT",
-                    "FUL",
-                    "ECHO",
-                    "RWT",
-                    "ITG",
-                    "AFG",
-                    "ASGN",
-                    "MPWR",
-                    "UDOW",
-                    "BGS",
-                    "ELP",
-                    "SPB",
-                    "MDCA",
-                    "CBZ",
-                    "IDCC",
-                    "CERS",
-                    "OMER",
-                    "CMO",
-                    "AAWW",
-                    "CBPX",
-                    "AVG",
-                    "FCN",
-                    "GDOT",
-                    "GWPH",
-                    "BHE",
-                    "NK",
-                    "RATE",
-                    "FMI",
-                    "CYT",
-                    "PBH",
-                    "GHDX",
-                    "LG",
-                    "UIS",
-                    "TM",
-                    "HCLP",
-                    "AXON",
-                    "TWI",
-                    "DLX",
-                    "HTLD",
-                    "HURN",
-                    "AFFX",
-                    "MATX",
-                    "RMTI",
-                    "CPST",
-                    "GHL",
-                    "AYR",
-                    "COR",
-                    "TTC",
-                    "RPT",
-                    "SNR",
-                    "HELE",
-                    "SPLV",
-                    "ITRI",
-                    "PSEC",
-                    "AAOI",
-                    "HCSG",
-                    "THRM",
-                    "PRI",
-                    "SQM",
-                    "OTEX",
-                    "BCOR",
-                    "ECPG",
-                    "APAM",
-                    "ENLC",
-                    "NDSN",
-                    "LYG",
-                    "QTM",
-                    "WGO",
-                    "HOV",
-                    "PVG",
-                    "XHR",
-                    "SYT",
-                    "MOMO",
-                    "TMV",
-                    "REXX",
-                    "XNCR",
-                    "THG",
-                    "CAMP",
-                    "CHUY",
-                    "SIR",
-                    "SWIR",
-                    "CURE",
-                    "QTS",
-                    "MR",
-                    "ARR",
-                    "AHL",
-                    "BABY",
-                    "WTI",
-                    "BPFH",
-                    "COT",
-                    "CSH",
-                    "GLF",
-                    "ENPH",
-                    "MEI",
-                    "RGS",
-                    "GSM",
-                    "NVCR",
-                    "SRNE",
-                    "CZZ",
-                    "ENI",
-                    "EFII",
-                    "NHTC",
-                    "FEIC",
-                    "SSP",
-                    "INGN",
-                    "RENT",
-                    "AINV",
-                    "PKY",
-                    "BDC",
-                    "TDY",
-                    "MTN",
-                    "ADTN",
-                    "LINE",
-                    "CORE",
-                    "CARA",
-                    "DLB",
-                    "PSO",
-                    "IRBT",
-                    "NG",
-                    "HQY",
-                    "TTS",
-                    "VNET",
-                    "COWN",
-                    "TUES",
-                    "AMDA",
-                    "PGTI",
-                    "IART",
-                    "TBI",
-                    "PRTY",
-                    "ABM",
-                    "NNBR",
-                    "LJPC",
-                    "NJR",
-                    "WBAI",
-                    "SIMO",
-                    "IRDM",
-                    "ARII",
-                    "ACTG",
-                    "WIBC",
-                    "MTRX",
-                    "MWA",
-                    "WRE",
-                    "CNW",
-                    "GCO",
-                    "VSTO",
-                    "KKD",
-                    "AKBA",
-                    "ENTG",
-                    "WWWW",
-                    "PEI",
-                    "PGH",
-                    "OTIC",
-                    "RDWR",
-                    "GGAL",
-                    "HLTH",
-                    "WD",
-                    "FELE",
-                    "NWBI",
-                    "VNR",
-                    "XLRN",
-                    "BSFT",
-                    "KPTI",
-                    "DST",
-                    "IYT",
-                    "VSAT",
-                    "VSTM",
-                    "BTG",
-                    "LNCO",
-                    "MDP",
-                    "BOKF",
-                    "NEP",
-                    "SMG",
-                    "IRET",
-                    "RUSS",
-                    "PRLB",
-                    "MRKT",
-                    "SIGM",
-                    "B",
-                    "HWAY",
-                    "CHRS",
-                    "APOG",
-                    "KYN",
-                    "AMRI",
-                    "KW",
-                    "VIG",
-                    "CMP",
-                    "GIB",
-                    "VTL",
-                    "BECN",
-                    "SJI",
-                    "EBIX",
-                    "CRAY",
-                    "LFC",
-                    "OZM",
-                    "BCO",
-                    "SOHU",
-                    "SLAB",
-                    "CMPR",
-                    "ORIG",
-                    "EWU",
-                    "EIDO",
-                    "CHE",
-                    "HMC",
-                    "MKSI",
-                    "SAIC",
-                    "ATRO",
-                    "PJT",
-                    "RBY",
-                    "ARCB",
-                    "QSII",
-                    "PLKI",
-                    "SAH",
-                    "BGC",
-                    "BLCM",
-                    "DENN",
-                    "STAR",
-                    "IMMU",
-                    "TTEK",
-                    "ROL",
-                    "TRU",
-                    "DIN",
-                    "DOOR",
-                    "USMV",
-                    "IVE",
-                    "FRO",
-                    "PUK",
-                    "OMCL",
-                    "GNMK",
-                    "FMSA",
-                    "FRPT",
-                    "UNVR",
-                    "RRTS",
-                    "NHI",
-                    "HRG",
-                    "LZB",
-                    "CEO",
-                    "CACC",
-                    "APDN",
-                    "RUBI",
-                    "EWP",
-                    "HABT",
-                    "MTX",
-                    "OIL",
-                    "BEAT",
-                    "MHO",
-                    "AEIS",
-                    "KFY",
-                    "HZO",
-                    "ACCO",
-                    "CSLT",
-                    "AIT",
-                    "CAL",
-                    "LOPE",
-                    "CATM",
-                    "ZOES",
-                    "BKH",
-                    "OCUL",
-                    "CSU",
-                    "PTX",
-                    "HTH",
-                    "OGS",
-                    "TU",
-                    "GNCA",
-                    "SKM",
-                    "MTGE",
-                    "CRH",
-                    "VRA",
-                    "NLS",
-                    "KCG",
-                    "QDEL",
-                    "CPPL",
-                    "OVTI",
-                    "CCG",
-                    "MORE",
-                    "CNCE",
-                    "NPBC",
-                    "ACET",
-                    "INOV",
-                    "STAY",
-                    "ISLE",
-                    "AMCC",
-                    "EWI",
-                    "AAC",
-                    "PMC",
-                    "BIN",
-                    "BND",
-                    "NWBO",
-                    "TSEM",
-                    "CANF",
-                    "HUBG",
-                    "SDY",
-                    "ALGT",
-                    "FARO",
-                    "AM",
-                    "BV",
-                    "TIVO",
-                    "USCR",
-                    "ABCB",
-                    "FLML",
-                    "FIG",
-                    "ATEN",
-                    "FWRD",
-                    "CSGP",
-                    "MCRB",
-                    "DORM",
-                    "ELY",
-                    "ATRC",
-                    "RHP",
-                    "SWX",
-                    "AAT",
-                    "CVEO",
-                    "PLXS",
-                    "FN",
-                    "NPTN",
-                    "PCH",
-                    "WMS",
-                    "SAM",
-                    "HOMB",
-                    "DRYS",
-                    "MINI",
-                    "EWM",
-                    "GVA",
-                    "QIWI",
-                    "LMNX",
-                    "VLRS",
-                    "CWEI",
-                    "BTI",
-                    "INN",
-                    "CTRV",
-                    "EGHT",
-                    "WAGE",
-                    "CENTA",
-                    "PRGS",
-                    "HF",
-                    "HHC",
-                    "GB",
-                    "IOC",
-                    "VIRT",
-                    "IBKC",
-                    "LTXB",
-                    "EMB",
-                    "QURE",
-                    "SLGN",
-                    "AVA",
-                    "SOXL",
-                    "YGE",
-                    "LXFT",
-                    "AJRD",
-                    "TRS",
-                    "MNRO",
-                    "ABAX",
-                    "TAST",
-                    "CLLS",
-                    "SCLN",
-                    "EZA",
-                    "YDLE",
-                    "SNN",
-                    "EEQ",
-                    "SONS",
-                    "WB",
-                    "NOAH",
-                    "SYKE",
-                    "EFOI",
-                    "AMWD",
-                    "GASL",
-                    "SA",
-                    "TPC",
-                    "SAIA",
-                    "FICO",
-                    "CFFN",
-                    "SHOR",
-                    "ROCK",
-                    "WING",
-                    "ALE",
-                    "VDE",
-                    "HUBS",
-                    "VHT",
-                    "SUN",
-                    "RCAP",
-                    "EXTR",
-                    "UCBI",
-                    "OCLR",
-                    "MITL",
-                    "OCAT",
-                    "PRAH",
-                    "BOH",
-                    "SIEN",
-                    "GEF",
-                    "ANH",
-                    "RUN",
-                    "FBP",
-                    "IWS",
-                    "UMC",
-                    "VRTU",
-                    "TSE",
-                    "CPL",
-                    "PACD",
-                    "XES",
-                    "AMPH",
-                    "ACXM",
-                    "FCG",
-                    "E",
-                    "CYNO",
-                    "LIVN",
-                    "FNFV",
-                    "JAKK",
-                    "NPO",
-                    "LPG",
-                    "MPAA",
-                    "PRA",
-                    "CKEC",
-                    "PCTY",
-                    "CHU",
-                    "ESL",
-                    "VYM",
-                    "CRK",
-                    "OIBR.C",
-                    "WGP",
-                    "TIP",
-                    "AMC",
-                    "PSXP",
-                    "RAIL",
-                    "AKR",
-                    "MSTR",
-                    "NEO",
-                    "NATI",
-                    "LE",
-                    "FCB",
-                    "BBW",
-                    "BKLN",
-                    "BOBE",
-                    "STM",
-                    "SWAY",
-                    "ICUI",
-                    "FSP",
-                    "CW",
-                    "KOP",
-                    "FV",
-                    "SWM",
-                    "WTS",
-                    "NEOG",
-                    "SODA",
-                    "RNET",
-                    "IDRA",
-                    "RESI",
-                    "MRTX",
-                    "FBT",
-                    "EVRI",
-                    "SPWH",
-                    "IO",
-                    "SMFG",
-                    "NM",
-                    "CCC",
-                    "BIOS",
-                    "OME",
-                    "MCHI",
-                    "BNED",
-                    "BRC",
-                    "MMI",
-                    "YANG",
-                    "ASX",
-                    "TXTR",
-                    "SHOP",
-                    "PPP",
-                    "ASC",
-                    "LBIO",
-                    "EVTC",
-                    "FCFS",
-                    "ALU",
-                    "CBB",
-                    "IMKTA",
-                    "LRN",
-                    "USNA",
-                    "FLXN",
-                    "VUG",
-                    "DBC",
-                    "RPRX",
-                    "IRC",
-                    "VB",
-                    "CHH",
-                    "AMBC",
-                    "LOGI",
-                    "COLB",
-                    "IAU",
-                    "SCHL",
-                    "WCIC",
-                    "ONTY",
-                    "AREX",
-                    "LFUS",
-                    "IYF",
-                    "JW.A",
-                    "BLKB",
-                    "NCS",
-                    "GOL",
-                    "SSNI",
-                    "TNP",
-                    "OMED",
-                    "AIR",
-                    "PSG",
-                    "OSIR",
-                    "CYBX",
-                    "UFPI",
-                    "ENOC",
-                    "DFRG",
-                    "IMPR",
-                    "EVC",
-                    "PRO",
-                    "VNCE",
-                    "PSMT",
-                    "LTC",
-                    "GEOS",
-                    "ENVA",
-                    "CLS",
-                    "GLOB",
-                    "DRN",
-                    "WMC",
-                    "WIT",
-                    "KRO",
-                    "ACRX",
-                    "NCI",
-                    "DBEU",
-                    "CNMD",
-                    "LBRDK",
-                    "GTT",
-                    "RSE",
-                    "BMA",
-                    "BFAM",
-                    "RAS",
-                    "TCS",
-                    "JASO",
-                    "RSO",
-                    "PLT",
-                    "AEG",
-                    "WX",
-                    "WSTC",
-                    "IPCM",
-                    "IBP",
-                    "ONVO",
-                    "PSQ",
-                    "THLD",
-                    "MYRG",
-                    "SHLM",
-                    "CBPO",
-                    "RGR",
-                    "GTE",
-                    "EYES",
-                    "CKP",
-                    "BOJA",
-                    "CWT",
-                    "SPHS",
-                    "IEV",
-                    "OXM",
-                    "SXT",
-                    "ENSG",
-                    "TISI",
-                    "UEPS",
-                    "FXE",
-                    "BDBD",
-                    "I",
-                    "TCP",
-                    "PRE",
-                    "IOSP",
-                    "UMBF",
-                    "SCAI",
-                    "BRSS",
-                    "KNDI",
-                    "SCHB",
-                    "SIGI",
-                    "SALT",
-                    "BGFV",
-                    "ACLS",
-                    "ACAT",
-                    "HASI",
-                    "ARP",
-                    "RXDX",
-                    "CHT",
-                    "IYH",
-                    "QUAD",
-                    "ARI",
-                    "CEB",
-                    "BBH",
-                    "TFSL",
-                    "CHGG",
-                    "TACO",
-                    "TESO",
-                    "CACI",
-                    "TROV",
-                    "WGBS",
-                    "NYMT",
-                    "BW",
-                    "GDP",
-                    "MEMP",
-                    "POWI",
-                    "TLMR",
-                    "NVDQ",
-                    "ASTI",
-                    "CTRE",
-                    "WIFI",
-                    "ESND",
-                    "SSE",
-                    "PFS",
-                    "OSUR",
-                    "IILG",
-                    "ILF",
-                    "EXLS",
-                    "AVID",
-                    "HEZU",
-                    "BEL",
-                    "NEWM",
-                    "ROG",
-                    "UTEK",
-                    "VTAE",
-                    "GNRT",
-                    "RPXC",
-                    "HI",
-                    "NMM",
-                    "SCSC",
-                    "PBPB",
-                    "ALEX",
-                    "WIX",
-                    "PKD",
-                    "CPF",
-                    "TTMI",
-                    "AQXP",
-                    "ZAGG",
-                    "HOLI",
-                    "ARPI",
-                    "KLIC",
-                    "CMN",
-                    "IPHS",
-                    "LITE",
-                    "MHLD",
-                    "OLLI",
-                    "ANDE",
-                    "ADRO",
-                    "BNCL",
-                    "AFMD",
-                    "HE",
-                    "FLWS",
-                    "PNFP",
-                    "CVT",
-                    "AZZ",
-                    "AXE",
-                    "SSD",
-                    "TPRE",
-                    "FSS",
-                    "INAP",
-                    "VMEM",
-                    "REGI",
-                    "ENT",
-                    "KTOS",
-                    "NCMI",
-                    "DBVT",
-                    "HAFC",
-                    "BRKL",
-                    "GNCMA",
-                    "SPIL",
-                    "CVGW",
-                    "TEGP",
-                    "BBSI",
-                    "HGR",
-                    "KTWO",
-                    "TBPH",
-                    "PRIM",
-                    "CSGS",
-                    "TOO",
-                    "NSIT",
-                    "BIP",
-                    "TOUR",
-                    "BRKS",
-                    "CTRN",
-                    "TGP",
-                    "FPO",
-                    "LPL",
-                    "HNI",
-                    "CGI",
-                    "WRLD",
-                    "EGP",
-                    "NTT",
-                    "CAFD",
-                    "RTEC",
-                    "RUSHA",
-                    "TLK",
-                    "ERA",
-                    "SAND",
-                    "KNL",
-                    "CATO",
-                    "CLDT",
-                    "ETH",
-                    "NEWR",
-                    "CVTI",
-                    "SMLP",
-                    "NTRA",
-                    "FSC",
-                    "VSAR",
-                    "CVRR",
-                    "EZCH",
-                    "RT",
-                    "EVEP",
-                    "FRSH",
-                    "KMPR",
-                    "ECOM",
-                    "SNSS",
-                    "SYA",
-                    "DXPE",
-                    "PLNT",
-                    "GRBK",
-                    "ARLP",
-                    "HRTG",
-                    "REXR",
-                    "AWR",
-                    "VIVO",
-                    "ITEK",
-                    "AFOP",
-                    "APU",
-                    "CQP",
-                    "MOV",
-                    "AMRN",
-                    "TNGO",
-                    "KT",
-                    "BLT",
-                    "SFS",
-                    "TA",
-                    "NX",
-                    "FTD",
-                    "DHX",
-                    "WPP",
-                    "DIOD",
-                    "KFRC",
-                    "MLI",
-                    "HCC",
-                    "EDE",
-                    "SQNM",
-                    "ABUS",
-                    "EUM",
-                    "TPLM",
-                    "BIO",
-                    "MCY",
-                    "AVD",
-                    "TEP",
-                    "BETR",
-                    "PTR",
-                    "RPD",
-                    "SFNC",
-                    "EPAY",
-                    "MSA",
-                    "LNN",
-                    "MLPI",
-                    "RBS",
-                    "FFIN",
-                    "LHCG",
-                    "BFR",
-                    "CNS",
-                    "COHR",
-                    "PAHC",
-                    "MCFT",
-                    "AEGN",
-                    "FFBC",
-                    "NICE",
-                    "THR",
-                    "NDRM",
-                    "JBT",
-                    "EGOV",
-                    "VO",
-                    "FCF",
-                    "FRED",
-                    "ARCO",
-                    "DMRC",
-                    "MTU",
-                    "CALX",
-                    "MCRN",
-                    "LANC",
-                    "TNDM",
-                    "LF",
-                    "IWR",
-                    "SFG",
-                    "AHP",
-                    "ABTL",
-                    "CKH",
-                    "CALA",
-                    "CSFL",
-                    "CTIC",
-                    "BPMC",
-                    "WAIR",
-                    "IBOC",
-                    "BBEP",
-                    "KRNY",
-                    "NGHC",
-                    "QTWO",
-                    "EWQ",
-                    "WABC",
-                    "TYG",
-                    "MUX",
-                    "ASTE",
-                    "FSTR",
-                    "RWX",
-                    "GNL",
-                    "FBC",
-                    "AI",
-                    "AIMT",
-                    "VRNS",
-                    "SYNT",
-                    "SGI",
-                    "HEFA",
-                    "TRNO",
-                    "IMMR",
-                    "KODK",
-                    "UVV",
-                    "SFXE",
-                    "ANIK",
-                    "BSAC",
-                    "BNFT",
-                    "ECYT",
-                    "CEVA",
-                    "SPTN",
-                    "REX",
-                    "ISCA",
-                    "ZIXI",
-                    "EE",
-                    "IG",
-                    "TX",
-                    "GMLP",
-                    "BKFS",
-                    "PEGA",
-                    "ICF",
-                    "VGT",
-                    "DW",
-                    "NBHC",
-                    "KONA",
-                    "UEIC",
-                    "LABL",
-                    "CDR",
-                    "GBT",
-                    "FDML",
-                    "UWM",
-                    "EJ",
-                    "PSB",
-                    "TUR",
-                    "URTY",
-                    "KANG",
-                    "SDLP",
-                    "BCPC",
-                    "ECOL",
-                    "IIVI",
-                    "CFMS",
-                    "RP",
-                    "LQDT",
-                    "JIVE",
-                    "OMG",
-                    "FSIC",
-                    "ADAP",
-                    "EUO",
-                    "FOR",
-                    "AIMC",
-                    "AVAV",
-                    "CEM",
-                    "RNST",
-                    "GKOS",
-                    "CPLA",
-                    "AMCN",
-                    "SGRY",
-                    "HCI",
-                    "FORM",
-                    "CBU",
-                    "SEMI",
-                    "SBY",
-                    "BANR",
-                    "DOG",
-                    "SCZ",
-                    "EZPW",
-                    "AGRO",
-                    "SPH",
-                    "FNBC",
-                    "CENT",
-                    "PLAB",
-                    "TOWR",
-                    "FDP",
-                    "PMTS",
-                    "TANH",
-                    "UNIS",
-                    "JMBA",
-                    "RAVN",
-                    "CNSL",
-                    "ANFI",
-                    "ORAN",
-                    "CPRX",
-                    "OSTK",
-                    "RFP",
-                    "VIIX",
-                    "PLUS",
-                    "PRSC",
-                    "CECO",
-                    "IHG",
-                    "TUBE",
-                    "CHMA",
-                    "LFL",
-                    "MANT",
-                    "KEYW",
-                    "PJC",
-                    "BLX",
-                    "HEWG",
-                    "PGI",
-                    "INTL",
-                    "LPSN",
-                    "LPCN",
-                    "SNP",
-                    "MESG",
-                    "CIVI",
-                    "FOGO",
-                    "CLMT",
-                    "SYN",
-                    "EWL",
-                    "MOG.A",
-                    "NSH",
-                    "XPH",
-                    "XCRA",
-                    "CTRL",
-                    "CCMP",
-                    "WAC",
-                    "TPUB",
-                    "HIFR",
-                    "MX",
-                    "SNMX",
-                    "ANGO",
-                    "CHEF",
-                    "UBSH",
-                    "SREV",
-                    "ININ",
-                    "AGII",
-                    "VSLR",
-                    "LOXO",
-                    "CAJ",
-                    "TYPE",
-                    "GLRE",
-                    "SCHX",
-                    "ONDK",
-                    "OAK",
-                    "NMIH",
-                    "EBSB",
-                    "RECN",
-                    "TEO",
-                    "TKC",
-                    "HLIT",
-                    "PIN",
-                    "DERM",
-                    "RUTH",
-                    "KEP",
-                    "SMP",
-                    "EXPO",
-                    "PETS",
-                    "CBF",
-                    "PAC",
-                    "CPSI",
-                    "AAON",
-                    "STRA",
-                    "MOBL",
-                    "NSP",
-                    "DM",
-                    "AXAS",
-                    "FLY",
-                    "MEET",
-                    "EGL",
-                    "FF",
-                    "MTRN",
-                    "RIGL",
-                    "TEDU",
-                    "RLD",
-                    "GTS",
-                    "AVX",
-                    "DANG",
-                    "RDNT",
-                    "MG",
-                    "RTK",
-                    "MB",
-                    "FUEL",
-                    "NRE",
-                    "HMN",
-                    "CUBI",
-                    "SSW",
-                    "SQI",
-                    "VT",
-                    "EEMV",
-                    "RLI",
-                    "FOMX",
-                    "YDKN",
-                    "IEZ",
-                    "GK",
-                    "TARO",
-                    "SHY",
-                    "MAIN",
-                    "HEI",
-                    "FXH",
-                    "CORT",
-                    "EVDY",
-                    "SATS",
-                    "SCHF",
-                    "ASR",
-                    "VPL",
-                    "COHU",
-                    "DUG",
-                    "VRTS",
-                    "ROLL",
-                    "VSCP",
-                    "MKL",
-                    "AGX",
-                    "PJP",
-                    "NEWP",
-                    "DCOM",
-                    "UIL",
-                    "NILE",
-                    "ARDX",
-                    "MFLX",
-                    "PDFS",
-                    "MDGN",
-                    "NSA",
-                    "RTIX",
-                    "IL",
-                    "VASC",
-                    "ARC",
-                    "PATK",
-                    "UBIO",
-                    "PSIX",
-                    "SHEN",
-                    "ALOG",
-                    "HTGC",
-                    "CLW",
-                    "IVC",
-                    "UHAL",
-                    "APEI",
-                    "ALRM",
-                    "MATW",
-                    "CIR",
-                    "BANC",
-                    "EVH",
-                    "CTCM",
-                    "CUB",
-                    "EFAV",
-                    "NVGS",
-                    "AMTG",
-                    "MITT",
-                    "NWN",
-                    "EPHE",
-                    "EIG",
-                    "SQBG",
-                    "CNA",
-                    "FMS",
-                    "UIHC",
-                    "EXAR",
-                    "MRTN",
-                    "TBBK",
-                    "AGTC",
-                    "AV",
-                    "OSIS",
-                    "ASCMA",
-                    "NIHD",
-                    "SCL",
-                    "WNS",
-                    "WIRE",
-                    "RSTI",
-                    "OFIX",
-                    "NTG",
-                    "KOF",
-                    "VLP",
-                    "IXYS",
-                    "HCKT",
-                    "ATRS",
-                    "SNAK",
-                    "HSTM",
-                    "KIRK",
-                    "CFNL",
-                    "INDB",
-                    "IDV",
-                    "GLDD",
-                    "ADMS",
-                    "HMST",
-                    "SFY",
-                    "EGBN",
-                    "PFGC",
-                    "AIN",
-                    "EGLT",
-                    "GENE",
-                    "PPHM",
-                    "GTY",
-                    "URE",
-                    "GALE",
-                    "Y",
-                    "LEJU",
-                    "IWP",
-                    "ORA",
-                    "SPSC",
-                    "BXE",
-                    "DSX",
-                    "EBF",
-                    "MNR",
-                    "NMFC",
-                    "NAVB",
-                    "BPT",
-                    "SID",
-                    "CWB",
-                    "SJNK",
-                    "FGP",
-                    "NBTB",
-                    "TRIB",
-                    "EHTH",
-                    "NADL",
-                    "NNI",
-                    "CMRE",
-                    "LBRDA",
-                    "BDE",
-                    "ATSG",
-                    "PN",
-                    "CBA",
-                    "VXUS",
-                    "ORIT",
-                    "BZUN",
-                    "NEOS",
-                    "SMRT",
-                    "BSV",
-                    "AMFW",
-                    "FIBK",
-                    "GLT",
-                    "MGRC",
-                    "HHS",
-                    "FGL",
-                    "MPG",
-                    "CORR",
-                    "HEP",
-                    "GI",
-                    "USLV",
-                    "SZYM",
-                    "STC",
-                    "MRCY",
-                    "IEO",
-                    "DRNA",
-                    "RWR",
-                    "GFF",
-                    "ORN",
-                    "MED",
-                    "AYA",
-                    "ORBC",
-                    "LILAK",
-                    "DTSI",
-                    "IMDZ",
-                    "BMI",
-                    "BT",
-                    "CAI",
-                    "USPH",
-                    "POWR",
-                    "NEU",
-                    "ACTA",
-                    "TKAI",
-                    "KB",
-                    "ALIM",
-                    "ADC",
-                    "AVH",
-                    "NP",
-                    "MMYT",
-                    "RRMS",
-                    "DAKT",
-                    "CMTL",
-                    "MTSN",
-                    "CRIS",
-                    "CPLP",
-                    "VEDL",
-                    "PAM",
-                    "AB",
-                    "PGX",
-                    "NSTG",
-                    "XONE",
-                    "CYOU",
-                    "UAM",
-                    "CSV",
-                    "THD",
-                    "SKUL",
-                    "JJSF",
-                    "SFBS",
-                    "AMSF",
-                    "TRST",
-                    "MIFI",
-                    "HY",
-                    "NVIV",
-                    "CHFC",
-                    "FRME",
-                    "JONE",
-                    "BSQR",
-                    "GORO",
-                    "VHC",
-                    "IWV",
-                    "CYTR",
-                    "KWR",
-                    "NBBC",
-                    "FCSC",
-                    "CGEN",
-                    "CECE",
-                    "UCTT",
-                    "OMAM",
-                    "MGI",
-                    "HACK",
-                    "SSB",
-                    "EWS",
-                    "MITK",
-                    "EXG",
-                    "LGCY",
-                    "CCXI",
-                    "NDLS",
-                    "PLOW",
-                    "JE",
-                    "DNAI",
-                    "BRZU",
-                    "UFI",
-                    "ADUS",
-                    "TITN",
-                    "UUP",
-                    "MBUU",
-                    "AXDX",
-                    "EQGP",
-                    "MC",
-                    "SCHA",
-                    "JOBS",
-                    "PNNT",
-                    "DX",
-                    "NVLS",
-                    "JBSS",
-                    "MYE",
-                    "EXK",
-                    "WG",
-                    "WSFS",
-                    "RNDY",
-                    "VCR",
-                    "IYW",
-                    "LBY",
-                    "ENBL",
-                    "MOD",
-                    "NML",
-                    "SCHD",
-                    "LION",
-                    "CEF",
-                    "DIG",
-                    "GST",
-                    "ESE",
-                    "WSR",
-                    "SLRC",
-                    "BSTC",
-                    "BTX",
-                    "BHLB",
-                    "SCVL",
-                    "CHAD",
-                    "WDFC",
-                    "NNA",
-                    "CDZI",
-                    "RGNX",
-                    "HTHT",
-                    "UNF",
-                    "NCT",
-                    "NQ",
-                    "RLGT",
-                    "MORN",
-                    "GCAP",
-                    "PSEM",
-                    "VVI",
-                    "APIC",
-                    "STBZ",
-                    "MCF",
-                    "GEN",
-                    "CTR",
-                    "NFBK",
-                    "CCIH",
-                    "LILA",
-                    "UNTD",
-                    "QUOT",
-                    "HEAR",
-                    "ERI",
-                    "ATNI",
-                    "RDEN",
-                    "AMOT",
-                    "ETRM",
-                    "HIVE",
-                    "FTAI",
-                    "BZQ",
-                    "ALDW",
-                    "CRVL",
-                    "VBLT",
-                    "UEC",
-                    "WSBC",
-                    "SXI",
-                    "GOOD",
-                    "OPWR",
-                    "DTEA",
-                    "BBOX",
-                    "KAMN",
-                    "SXCP",
-                    "PFSW",
-                    "TG",
-                    "KMF",
-                    "ORC",
-                    "COKE",
-                    "NVTA",
-                    "ACWX",
-                    "LDL",
-                    "ECH",
-                    "EXLP",
-                    "LADR",
-                    "MERC",
-                    "QNST",
-                    "GLPG",
-                    "HYT",
-                    "IDU",
-                    "ACRE",
-                    "AFAM",
-                    "ZEUS",
-                    "TOWN",
-                    "ORBK",
-                    "CNAT",
-                    "PRMW",
-                    "BSET",
-                    "PRFT",
-                    "FXG",
-                    "CALL",
-                    "GNE",
-                    "MTSC",
-                    "COLL",
-                    "LMAT",
-                    "PGND",
-                    "FOXF",
-                    "MXWL",
-                    "VCRA",
-                    "USAC",
-                    "TIER",
-                    "CYTX",
-                    "IMOS",
-                    "KYE",
-                    "CHDN",
-                    "HLI",
-                    "SRG",
-                    "EOC",
-                    "LIOX",
-                    "AEPI",
-                    "OGXI",
-                    "PGEM",
-                    "GMZ",
-                    "UBA",
-                    "USM",
-                    "CHAU",
-                    "VOE",
-                    "KAI",
-                    "WBK",
-                    "EPP",
-                    "PHI",
-                    "VCSH",
-                    "DRRX",
-                    "VXZ",
-                    "HBM",
-                    "STBA",
-                    "VXF",
-                    "CCS",
-                    "VFH",
-                    "PERY",
-                    "NMR",
-                    "CPK",
-                    "VET",
-                    "ETY",
-                    "DTLK",
-                    "VV",
-                    "SBSI",
-                    "FNHC",
-                    "CTS",
-                    "UQM",
-                    "PKOH",
-                    "CCU",
-                    "ISSI",
-                    "TTEC",
-                    "HNRG",
-                    "PCI",
-                    "EMKR",
-                    "ALG",
-                    "PFNX",
-                    "GLP",
-                    "FCPT",
-                    "TNC",
-                    "OMN",
-                    "VRNG",
-                    "DSCI",
-                    "SBLK",
-                    "HVT",
-                    "IBCP",
-                    "WPPGY",
-                    "IDT",
-                    "VBR",
-                    "PARR",
-                    "TMF",
-                    "PEN",
-                    "BBRG",
-                    "BKCC",
-                    "GPX",
-                    "SB",
-                    "VEC",
-                    "BSM",
-                    "UAN",
-                    "TAC",
-                    "RENX",
-                    "PCY",
-                    "CMCO",
-                    "EDAP",
-                    "SUP",
-                    "EGY",
-                    "NES",
-                    "EFSC",
-                    "SNHY",
-                    "FMO",
-                    "CQH",
-                    "HFWA",
-                    "FUR",
-                    "GER",
-                    "DEM",
-                    "PXLW",
-                    "SPOK",
-                    "NANO",
-                    "AT",
-                    "FEI",
-                    "CBK",
-                    "IJK",
-                    "MPSX",
-                    "DBJP",
-                    "HAYN",
-                    "MSTX",
-                    "IJT",
-                    "IRT",
-                    "CVCO",
-                    "GREK",
-                    "MY",
-                    "CFI",
-                    "STAA",
-                    "KMG",
-                    "STNR",
-                    "PPBI",
-                    "MGEE",
-                    "FIVN",
-                    "CPS",
-                    "TCPC",
-                    "FIZZ",
-                    "EMO",
-                    "PICO",
-                    "SPNE",
-                    "MBB",
-                    "GLOP",
-                    "DGII",
-                    "INWK",
-                    "ZYNE",
-                    "BIOD",
-                    "UBNK",
-                    "TCAP",
-                    "FXN",
-                    "GIFI",
-                    "FSM",
-                    "LMOS",
-                    "BNCN",
-                    "MMLP",
-                    "HCHC",
-                    "VRTV",
-                    "ICFI",
-                    "BOOM",
-                    "HSKA",
-                    "IPAR",
-                    "TRXC",
-                    "EOX",
-                    "TR",
-                    "AROC",
-                    "BFS",
-                    "BLDP",
-                    "TCX",
-                    "RXII",
-                    "EVV",
-                    "NRP",
-                    "SNOW",
-                    "SOXS",
-                    "BCOV",
-                    "JRVR",
-                    "PNX",
-                    "HQH",
-                    "COBZ",
-                    "HDV",
-                    "JMF",
-                    "AHH",
-                    "ETM",
-                    "LAYN",
-                    "JMG",
-                    "AGQ",
-                    "LOB",
-                    "ENZ",
-                    "DCM",
-                    "FVE",
-                    "XOOM",
-                    "AVOL",
-                    "VPU",
-                    "WPRT",
-                    "IPCC",
-                    "DSU",
-                    "OPB",
-                    "HMIN",
-                    "HEI.A",
-                    "MEP",
-                    "STON",
-                    "MCC",
-                    "OTTR",
-                    "ADHD",
-                    "IHI",
-                    "VOXX",
-                    "CIU",
-                    "TAXI",
-                    "ESI",
-                    "PKE",
-                    "POWL",
-                    "THOR",
-                    "MEIP",
-                    "EXTN",
-                    "MFG",
-                    "VNM",
-                    "REM",
-                    "WLKP",
-                    "NWPX",
-                    "CLM",
-                    "FRP",
-                    "NGS",
-                    "NEWT",
-                    "LYTS",
-                    "VCIT",
-                    "FUN",
-                    "LLNW",
-                    "CORI",
-                    "TNAV",
-                    "WNRL",
-                    "LIFE",
-                    "GOGL",
-                    "REMY",
-                    "RM",
-                    "LKFN",
-                    "LAKE",
-                    "LORL",
-                    "VNQI",
-                    "CAF",
-                    "FATE",
-                    "HSII",
-                    "HTBK",
-                    "SEAC",
-                    "TTOO",
-                    "UDF",
-                    "GBDC",
-                    "FENX",
-                    "HART",
-                    "EPIQ",
-                    "SCHG",
-                    "TCBK",
-                    "COUP",
-                    "IHF",
-                    "CCO",
-                    "TC",
-                    "RELX",
-                    "KELYA",
-                    "CNNX",
-                    "MIE",
-                    "TSLX",
-                    "PFBC",
-                    "ACW",
-                    "JQC",
-                    "KBAL",
-                    "NTLS",
-                    "CRMD",
-                    "SBCF",
-                    "TSYS",
-                    "CARB",
-                    "ASMB",
-                    "RMAX",
-                    "SAFT",
-                    "RIGP",
-                    "NTK",
-                    "IEI",
-                    "NYMX",
-                    "SHG",
-                    "DSL",
-                    "UHT",
-                    "FFIC",
-                    "SCHE",
-                    "FRM",
-                    "NGVC",
-                    "RYI",
-                    "EPOL",
-                    "ELOS",
-                    "VCYT",
-                    "TLYS",
-                    "SRDX",
-                    "APPF",
-                    "BLPH",
-                    "CTT",
-                    "GIG",
-                    "DL",
-                    "XTLY",
-                    "CHKE",
-                    "CSJ",
-                    "XOXO",
-                    "OHRP",
-                    "WRES",
-                    "NAO",
-                    "UFCS",
-                    "PRTK",
-                    "CNOB",
-                    "GHC",
-                    "FXD",
-                    "NPK",
-                    "HEWJ",
-                    "FFG",
-                    "MN",
-                    "AMID",
-                    "LDR",
-                    "CWST",
-                    "REI",
-                    "IIIN",
-                    "HME",
-                    "AGFS",
-                    "UTL",
-                    "MPO",
-                    "NORD",
-                    "VIMC",
-                    "IHE",
-                    "BCH",
-                    "BEBE",
-                    "VIA",
-                    "JAX",
-                    "IXC",
-                    "CAPL",
-                    "BRG",
-                    "CSBK",
-                    "MSO",
-                    "SCHM",
-                    "DEL",
-                    "APTS",
-                    "CIA",
-                    "RVT",
-                    "IYM",
-                    "CDI",
-                    "CRWN",
-                    "OPY",
-                    "MCS",
-                    "PRF",
-                    "CLMS",
-                    "TNXP",
-                    "RYH",
-                    "ANAT",
-                    "IRMD",
-                    "CVGI",
-                    "AMPE",
-                    "SRI",
-                    "NFJ",
-                    "USAK",
-                    "IGV",
-                    "FORR",
-                    "GSBD",
-                    "AVAL",
-                    "EDGE",
-                    "ASEI",
-                    "SJT",
-                    "LSG",
-                    "NAVG",
-                    "ABEO",
-                    "TBF",
-                    "SMED",
-                    "GGN",
-                    "SCPB",
-                    "ANW",
-                    "RELY",
-                    "ONE",
-                    "PRQR",
-                    "WMK",
-                    "CHCO",
-                    "WK",
-                    "MHG",
-                    "TZOO",
-                    "APPS",
-                    "SOL",
-                    "TRR",
-                    "FLXS",
-                    "ITOT",
-                    "BMTC",
-                    "CETV",
-                    "DEA",
-                    "INDY",
-                    "BPI",
-                    "EBR",
-                    "AAV",
-                    "PEBO",
-                    "NEON",
-                    "BELFB",
-                    "IEP",
-                    "OB",
-                    "STML",
-                    "DQ",
-                    "BGY",
-                    "MVIS",
-                    "CPSS",
-                    "RENN",
-                    "HZN",
-                    "FELP",
-                    "VDC",
-                    "WSBF",
-                    "PHIIK",
-                    "PID",
-                    "AHGP",
-                    "AOI",
-                    "REN",
-                    "NXRT",
-                    "TICC",
-                    "REV",
-                    "METR",
-                    "BHBK",
-                    "SCHH",
-                    "PHX",
-                    "ETW",
-                    "PLPM",
-                    "KMPH",
-                    "JNP",
-                    "SCHZ",
-                    "VLTC",
-                    "EVLV",
-                    "TLGT",
-                    "HTCH",
-                    "VBK",
-                    "AMBR",
-                    "HGG",
-                    "FWP",
-                    "SXE",
-                    "EVOK",
-                    "DCO",
-                    "TTGT",
-                    "PCCC",
-                    "FLKS",
-                    "RPG",
-                    "EXAC",
-                    "BKMU",
-                    "UTI",
-                    "CGIX",
-                    "FARM",
-                    "MCRI",
-                    "AMRS",
-                    "BNDX",
-                    "GBL",
-                    "MRNS",
-                    "MNI",
-                    "DJP",
-                    "MCEP",
-                    "ZNH",
-                    "VOT",
-                    "AST",
-                    "FM",
-                    "DEST",
-                    "SPNS",
-                    "SJW",
-                    "KNOP",
-                    "VVR",
-                    "GAME",
-                    "CRMT",
-                    "MGPI",
-                    "GSBC",
-                    "ENTL",
-                    "GUID",
-                    "PPT",
-                    "CABO",
-                    "AWP",
-                    "ZBIO",
-                    "AMU",
-                    "ALLT",
-                    "WATT",
-                    "AFH",
-                    "BONA",
-                    "IBTX",
-                    "MODN",
-                    "CWCO",
-                    "UTF",
-                    "QADA",
-                    "XBIT",
-                    "MTUM",
-                    "EVA",
-                    "GSVC",
-                    "QLTI",
-                    "BDCL",
-                    "MLPL",
-                    "HTBI",
-                    "CISG",
-                    "BDJ",
-                    "DSKY",
-                    "GNBC",
-                    "WTM",
-                    "FSFR",
-                    "DNN",
-                    "RQI",
-                    "GUNR",
-                    "CLFD",
-                    "HWCC",
-                    "LNTH",
-                    "OMAB",
-                    "HOFT",
-                    "KOPN",
-                    "CVM",
-                    "FSYS",
-                    "PLNR",
-                    "DXYN",
-                    "UNXL",
-                    "OMEX",
-                    "TCRD",
-                    "TBK",
-                    "ERIE",
-                    "LTS",
-                    "HYS",
-                    "VDTH",
-                    "SVA",
-                    "FDEF",
-                    "FAX",
-                    "SMM",
-                    "NETE",
-                    "BCX",
-                    "LINC",
-                    "CODI",
-                    "VTTI",
-                    "IYZ",
-                    "IMH",
-                    "FREE",
-                    "TIS",
-                    "EOD",
-                    "TRUP",
-                    "HIO",
-                    "BPY",
-                    "AZUR",
-                    "DSCO",
-                    "PCO",
-                    "IAT",
-                    "RMP",
-                    "DXLG",
-                    "TGD",
-                    "GHY",
-                    "DNP",
-                    "XUE",
-                    "DPG",
-                    "PDVW",
-                    "AVEO",
-                    "GIM",
-                    "YUME",
-                    "HTLF",
-                    "OEC",
-                    "EHIC",
-                    "SRS",
-                    "PQ",
-                    "TGA",
-                    "CRAI",
-                    "KCAP",
-                    "HIX",
-                    "ABCW",
-                    "SGM",
-                    "MUB",
-                    "PPH",
-                    "ASPN",
-                    "IRS",
-                    "RWO",
-                    "BGB",
-                    "CRESY",
-                    "HNP",
-                    "NCZ",
-                    "SPF",
-                    "CUTR",
-                    "ZPIN",
-                    "PHK",
-                    "LNDC",
-                    "HBIO",
-                    "RPV",
-                    "EBR.B",
-                    "BUSE",
-                    "LAWS",
-                    "FENG",
-                    "JO",
-                    "DTK",
-                    "RDI",
-                    "VTA",
-                    "PFSI",
-                    "CMLP",
-                    "DKT",
-                    "CSWC",
-                    "XIN",
-                    "NEOT",
-                    "AOD",
-                    "CBR",
-                    "SCON",
-                    "PZN",
-                    "PBE",
-                    "NUTR",
-                    "GAM",
-                    "MTL",
-                    "GEVO",
-                    "PZE",
-                    "CRCM",
-                    "DGLY",
-                    "FXP",
-                    "JGW",
-                    "FONR",
-                    "UUUU",
-                    "CUR",
-                    "FWM",
-                    "PFLT",
-                    "BOE",
-                    "LUX",
-                    "SPA",
-                    "BRK.A",
-                    "EDD",
-                    "TRIL",
-                    "XNET",
-                    "GPL",
-                    "TAN",
-                    "WYY",
-                    "CHA",
-                    "ATTU",
-                    "IGD",
-                    "ICA",
-                    "CEN",
-                    "DBA",
-                    "PTSI",
-                    "KE",
-                    "IFN",
-                    "ACRS",
-                    "TCON",
-                    "RNN",
-                    "PDP",
-                    "GRC",
-                    "THFF",
-                    "EFV",
-                    "ATNM",
-                    "XGTI",
-                    "IJS",
-                    "AVGR",
-                    "RNF",
-                    "LMNR",
-                    "FCNCA",
-                    "EOS",
-                    "RYT",
-                    "BONT",
-                    "GHM",
-                    "TI",
-                    "PPR",
-                    "CCLP",
-                    "AGM",
-                    "BGR",
-                    "SFE",
-                    "QUAL",
-                    "NHF",
-                    "PLX",
-                    "CVRS",
-                    "ISD",
-                    "CSWI",
-                    "FIF",
-                    "DWX",
-                    "HQL",
-                    "HCCI",
-                    "JFR",
-                    "CASH",
-                    "FPL",
-                    "PBFX",
-                    "CTWS",
-                    "SIL",
-                    "MRIN",
-                    "SOJA",
-                    "PKW",
-                    "STCK",
-                    "EDN",
-                    "CASS",
-                    "CPTA",
-                    "TRK",
-                    "SHV",
-                    "XSD",
-                    "WHG",
-                    "MHGC",
-                    "USAT",
-                    "ANY",
-                    "AXN",
-                    "UYG",
-                    "STFC",
-                    "SYBT",
-                    "TECL",
-                    "ACG",
-                    "PGF",
-                    "CBMG",
-                    "RMT",
-                    "IMMY",
-                    "DBO",
-                    "NSU",
-                    "NCIT",
-                    "MORL",
-                    "FEN",
-                    "IJJ",
-                    "TTHI",
-                    "CRY",
-                    "CSS",
-                    "THQ",
-                    "WASH",
-                    "DEG",
-                    "MGK",
-                    "TAX",
-                    "STEM",
-                    "AIQ",
-                    "SHLO",
-                    "DZZ",
-                    "PTXP",
-                    "OPXA",
-                    "IYC",
-                    "TBNK",
-                    "OLP",
-                    "NCV",
-                    "EWN",
-                    "VIS",
-                    "VAW",
-                    "EPRS",
-                    "KIE",
-                    "EPM",
-                    "MATR",
-                    "ONCS",
-                    "VGGL",
-                    "CSCD",
-                    "CTBI",
-                    "FXCM",
-                    "PER",
-                    "AOSL",
-                    "EAD",
-                    "CSQ",
-                    "CERE",
-                    "GSS",
-                    "NVEE",
-                    "ETV",
-                    "RCKY",
-                    "PIRS",
-                    "ERUS",
-                    "MXPT",
-                    "GSIG",
-                    "EBIO",
-                    "GNK",
-                    "SRCE",
-                    "GAIN",
-                    "ENZN",
-                    "BXMX",
-                    "NDP",
-                    "TGS",
-                    "NWHM",
-                    "PHYS",
-                    "ESIO",
-                    "AVHI",
-                    "EMLC",
-                    "CHW",
-                    "CDTX",
-                    "UACL",
-                    "IGR",
-                    "HWKN",
-                    "UCFC",
-                    "AFT",
-                    "MM",
-                    "THW",
-                    "IXJ",
-                    "FHLC",
-                    "TSQ",
-                    "DATE",
-                    "ONCY",
-                    "ABR",
-                    "GFA",
-                    "BIOA",
-                    "CACQ",
-                    "MNGA",
-                    "PSTB",
-                    "MLR",
-                    "SHOS",
-                    "USA",
-                    "CHCT",
-                    "GAB",
-                    "AVK",
-                    "IMN",
-                    "CTSO",
-                    "SASR",
-                    "LBAI",
-                    "SPRT",
-                    "CAS",
-                    "SP",
-                    "VSEC",
-                    "TWIN",
-                    "ASND",
-                    "CASI",
-                    "ICL",
-                    "CII",
-                    "GXC",
-                    "IBA",
-                    "GALT",
-                    "TLP",
-                    "HELI",
-                    "NOR",
-                    "BREW",
-                    "SPEX",
-                    "CHI",
-                    "PGC",
-                    "XXII",
-                    "CRNT",
-                    "TSC",
-                    "CNXR",
-                    "CLBS",
-                    "FPF",
-                    "EPU",
-                    "FLOT",
-                    "BIV",
-                    "MLNK",
-                    "OSBC",
-                    "JPS",
-                    "MSEX",
-                    "MLPN",
-                    "UPIP",
-                    "KEM",
-                    "SPCB",
-                    "AMRC",
-                    "FSD",
-                    "FBRC",
-                    "CEA",
-                    "TTP",
-                    "USAP",
-                    "TFI",
-                    "AEMD",
-                    "FXL",
-                    "IYK",
-                    "MNTX",
-                    "PDI",
-                    "BOND",
-                    "BANF",
-                    "EWD",
-                    "XPP",
-                    "ATLS",
-                    "VTVT",
-                    "SLTD",
-                    "FRA",
-                    "ARQL",
-                    "MRLN",
-                    "MOO",
-                    "SENEA",
-                    "QCRH",
-                    "CIGI",
-                    "REIS",
-                    "OAKS",
-                    "OCFC",
-                    "MCUR",
-                    "PICK",
-                    "ARTNA",
-                    "GNR",
-                    "OCIP",
-                    "JRO",
-                    "GDV",
-                    "SUBK",
-                    "ESD",
-                    "TPVG",
-                    "CLDN",
-                    "SBIO",
-                    "AGYS",
-                    "DKL",
-                    "IVTY",
-                    "RYL",
-                    "ECON",
-                    "PSLV",
-                    "DSE",
-                    "TBRA",
-                    "SCHV",
-                    "LFVN",
-                    "LEN.B",
-                    "CEFL",
-                    "OXFD",
-                    "GUSH",
-                    "VLGEA",
-                    "NHC",
-                    "DHY",
-                    "DHF",
-                    "PEO",
-                    "NAP",
-                    "BNK",
-                    "AXGN",
-                    "BDGE",
-                    "RNP",
-                    "VSS",
-                    "KED",
-                    "VICL",
-                    "SGBK",
-                    "CO",
-                    "HNH",
-                    "FLIC",
-                    "VPG",
-                    "CVO",
-                    "MCHX",
-                    "TRC",
-                    "NSSC",
-                    "NC",
-                    "ATHX",
-                    "MSFG",
-                    "BCA",
-                    "IXUS",
-                    "DXM",
-                    "XRM",
-                    "VICR",
-                    "FC",
-                    "AKAO",
-                    "RIC",
-                    "NEFF",
-                    "UMH",
-                    "CHY",
-                    "ETG",
-                    "NBY",
-                    "ESGR",
-                    "NIO",
-                    "PRKR",
-                    "GLAD",
-                    "FFC",
-                    "EMCI",
-                    "EFC",
-                    "COSI",
-                    "EVT",
-                    "ISR",
-                    "EYEG",
-                    "MIN",
-                    "MSB",
-                    "RDHL",
-                    "FCBC",
-                    "LDP",
-                    "SZMK",
-                    "ROYT",
-                    "PIM",
-                    "ZIV",
-                    "PTY",
-                    "CCD",
-                    "ACH",
-                    "IYG",
-                    "MINT",
-                    "KZ",
-                    "FXEN",
-                    "TMP",
-                    "GPP",
-                    "NSL",
-                    "CLY",
-                    "QQQX",
-                    "PRK",
-                    "PBMD",
-                    "DSKX",
-                    "BIT",
-                    "LCUT",
-                    "OKSB",
-                    "TAT",
-                    "RST",
-                    "EIM",
-                    "DGS",
-                    "AWF",
-                    "DFE",
-                    "MBWM",
-                    "ADX",
-                    "JPC",
-                    "DSPG",
-                    "ANAD",
-                    "GBNK",
-                    "NJ",
-                    "RCS",
-                    "RTH",
-                    "TESS",
-                    "AGRX",
-                    "ETJ",
-                    "FSB",
-                    "STK",
-                    "PHB",
-                    "SCIF",
-                    "HURC",
-                    "TREC",
-                    "AKO.B",
-                    "CLV",
-                    "HIL",
-                    "DLNG",
-                    "INDL",
-                    "BWG",
-                    "ITA",
-                    "RLH",
-                    "UGLD",
-                    "ELRC",
-                    "EARN",
-                    "EEV",
-                    "CATB",
-                    "VNOM",
-                    "IUSG",
-                    "RAVE",
-                    "NVEC",
-                    "HQCL",
-                    "FBIZ",
-                    "GXG",
-                    "SYUT",
-                    "BIL",
-                    "GLO",
-                    "FISI",
-                    "KCLI",
-                    "RGSE",
-                    "SGC",
-                    "PSUN",
-                    "ERN",
-                    "LWAY",
-                    "SHM",
-                    "PBT",
-                    "CMT",
-                    "GYLD",
-                    "EMLP",
-                    "EVF",
-                    "CDTI",
-                    "MZOR",
-                    "VJET",
-                    "CLSN",
-                    "IMUC",
-                    "OCIR",
-                    "RNWK",
-                    "HGT",
-                    "DSGX",
-                    "FENY",
-                    "CBYL",
-                    "IFON",
-                    "CBAY",
-                    "SPKE",
-                    "SCD",
-                    "ASHS",
-                    "NATL",
-                    "BOIL",
-                    "BEP",
-                    "WMAR",
-                    "OOMA",
-                    "SLP",
-                    "NRO",
-                    "STV",
-                    "UVSP",
-                    "ARES",
-                    "VMO",
-                    "RUSHB",
-                    "GWL",
-                    "BAA",
-                    "SBR",
-                    "NATH",
-                    "TY",
-                    "EVOL",
-                    "WIW",
-                    "LIND",
-                    "GABC",
-                    "EFT",
-                    "DRIP",
-                    "BH",
-                    "RBCAA",
-                    "ALSK",
-                    "GLPW",
-                    "SNC",
-                    "PTH",
-                    "TDF",
-                    "VKI",
-                    "HBNC",
-                    "COVS",
-                    "GBLI",
-                    "BTZ",
-                    "LXFR",
-                    "CNXC",
-                    "EFR",
-                    "CNXT",
-                    "ISH",
-                    "VGM",
-                    "ERC",
-                    "PMM",
-                    "DMD",
-                    "EFG",
-                    "BLV",
-                    "HCOM",
-                    "EEMS",
-                    "BBN",
-                    "KBWB",
-                    "BNSO",
-                    "PLG",
-                    "AGC",
-                    "CTMX",
-                    "FBNK",
-                    "CUNB",
-                    "BWX",
-                    "VUZI",
-                    "BST",
-                    "DPRX",
-                    "FBNC",
-                    "INF",
-                    "SRCLP",
-                    "AXPW",
-                    "WIP",
-                    "PFN",
-                    "JPI",
-                    "FTA",
-                    "NEWS",
-                    "DMLP",
-                    "FPX",
-                    "BIOC",
-                    "IX",
-                    "FXO",
-                    "KVHI",
-                    "SGNL",
-                    "WLDN",
-                    "AMSWA",
-                    "FXY",
-                    "STB",
-                    "TECS",
-                    "AIF",
-                    "HIE",
-                    "SWNC",
-                    "USDP",
-                    "ANCX",
-                    "PERI",
-                    "OXGN",
-                    "STRT",
-                    "FEX",
-                    "KTF",
-                    "CACB",
-                    "IYJ",
-                    "EHI",
-                    "CHFN",
-                    "TEI",
-                    "NTP",
-                    "PCBK",
-                    "BIND",
-                    "ZF",
-                    "PSK",
-                    "DCTH",
-                    "IGF",
-                    "SDIV",
-                    "BBLU",
-                    "SILC",
-                    "GOMO",
-                    "BTO",
-                    "IRR",
-                    "ELD",
-                    "EMG",
-                    "MVG",
-                    "BFIN",
-                    "GNT",
-                    "SYRX",
-                    "CYD",
-                    "CRBP",
-                    "DWSN",
-                    "AUDC",
-                    "NPM",
-                    "NAC",
-                    "BMRC",
-                    "PDT",
-                    "DAVE",
-                    "CYNA",
-                    "PTN",
-                    "CLCD",
-                    "REZ",
-                    "PCYG",
-                    "IVH",
-                    "CMFN",
-                    "CFRX",
-                    "HYF",
-                    "BIOL",
-                    "EWK",
-                    "CNCO",
-                    "WPCS",
-                    "QTWW",
-                    "DMTX",
-                    "NUV",
-                    "ATTO",
-                    "ROX",
-                    "ATEC",
-                    "ICD",
-                    "GSG",
-                    "SEB",
-                    "CUI",
-                    "OXLC",
-                    "CHKR",
-                    "VCLT",
-                    "NPP",
-                    "ASYS",
-                    "FDEU",
-                    "DBEM",
-                    "EWX",
-                    "ABDC",
-                    "CPXX",
-                    "JRJC",
-                    "SCM",
-                    "REPH",
-                    "SHYG",
-                    "SPHD",
-                    "IQI",
-                    "UCP",
-                    "GMAN",
-                    "MUI",
-                    "HTD",
-                    "BGH",
-                    "MUC",
-                    "ETB",
-                    "RITT",
-                    "MRGE",
-                    "WHF",
-                    "PTF",
-                    "DTO",
-                    "EPV",
-                    "STRL",
-                    "TRX",
-                    "VHI",
-                    "NAD",
-                    "VKQ",
-                    "MIL",
-                    "GSL",
-                    "NDRO",
-                    "PRCP",
-                    "AHC",
-                    "BLW",
-                    "ACWV",
-                    "URG",
-                    "ABCD",
-                    "CAE",
-                    "BBEPP",
-                    "YORW",
-                    "DIAX",
-                    "DRAD",
-                    "QUIK",
-                    "PVCT",
-                    "CTY",
-                    "CEL",
-                    "PSL",
-                    "CWI",
-                    "GVI",
-                    "PKO",
-                    "ASA",
-                    "GTIM",
-                    "VTWO",
-                    "BTT",
-                    "PGP",
-                    "REXI",
-                    "VNRCP",
-                    "HTM",
-                    "LEE",
-                    "CNV",
-                    "SNBC",
-                    "BGT",
-                    "EOI",
-                    "NOBL",
-                    "MCZ",
-                    "TDTT",
-                    "MVV",
-                    "PHD",
-                    "SCHO",
-                    "CLIR",
-                    "HYI",
-                    "STN",
-                    "AKG",
-                    "LMLP",
-                    "OHGI",
-                    "BLRX",
-                    "RFI",
-                    "RBCN",
-                    "ALCO",
-                    "CTZ",
-                    "UTG",
-                    "MRCC",
-                    "IDX",
-                    "IDE",
-                    "DMO",
-                    "AAMC",
-                    "CRD.B",
-                    "NAME",
-                    "ENZY",
-                    "JCE",
-                    "FTRPR",
-                    "JPEP",
-                    "SPDC",
-                    "NPI",
-                    "ALQA",
-                    "PLM",
-                    "BF.A",
-                    "DBL",
-                    "MMD",
-                    "CIK",
-                    "APDNW",
-                    "HEB",
-                    "SKF",
-                    "IWC",
-                    "BITI",
-                    "IGM",
-                    "NKSH",
-                    "VTIP",
-                    "TIPT",
-                    "VOX",
-                    "URA",
-                    "NMRX",
-                    "MYI",
-                    "OIA",
-                    "DXB",
-                    "IFGL",
-                    "FBIO",
-                    "HRZN",
-                    "FORD",
-                    "FXR",
-                    "QAI",
-                    "KBIO",
-                    "MLAB",
-                    "CTG",
-                    "MANU",
-                    "CBNJ",
-                    "MMT",
-                    "JMLP",
-                    "HYD",
-                    "EDF",
-                    "SDR",
-                    "NIE",
-                    "MFL",
-                    "CIDM",
-                    "PEK",
-                    "HSEB",
-                    "HNR",
-                    "TEAR",
-                    "SPAR",
-                    "CCBG",
-                    "ZSL",
-                    "ORMP",
-                    "IXN",
-                    "JOUT",
-                    "ICLD",
-                    "AROW",
-                    "ATRI",
-                    "HMLP",
-                    "CHLN",
-                    "ESCA",
-                    "BHK",
-                    "SCHC",
-                    "JMI",
-                    "FDUS",
-                    "EMD",
-                    "RSYS",
-                    "MRVC",
-                    "CNIT",
-                    "FFWM",
-                    "MVC",
-                    "AVNW",
-                    "MVF",
-                    "ICAD",
-                    "NEPT",
-                    "ACV",
-                    "HTY",
-                    "GLQ",
-                    "CELP",
-                    "TOTL",
-                    "UGA",
-                    "JRI",
-                    "DCA",
-                    "EDV",
-                    "IGA",
-                    "MGU",
-                    "GRAM",
-                    "HOTR",
-                    "DSM",
-                    "BHL",
-                    "CAC",
-                    "EXA",
-                    "IMNP",
-                    "ARDC",
-                    "NHS",
-                    "EVG",
-                    "INB",
-                    "SYX",
-                    "RXL",
-                    "SRLP",
-                    "PCOM",
-                    "CLCT",
-                    "CNTY",
-                    "LPTN",
-                    "DSLV",
-                    "FSV",
-                    "NYNY",
-                    "BWINB",
-                    "SRV",
-                    "GMF",
-                    "JGH",
-                    "MCN",
-                    "PMO",
-                    "BCLI",
-                    "ABTX",
-                    "TWER",
-                    "ACP",
-                    "HMTV",
-                    "DLN",
-                    "HYEM",
-                    "CERU",
-                    "PCN",
-                    "PXF",
-                    "PCTI",
-                    "PHO",
-                    "SYNL",
-                    "FXCB",
-                    "TRAK",
-                    "NMA",
-                    "MYD",
-                    "JTP",
-                    "PRGX",
-                    "YZC",
-                    "MOFG",
-                    "AFCO",
-                    "SDT",
-                    "SYPR",
-                    "HGH",
-                    "ZFC",
-                    "SURG",
-                    "JRS",
-                    "NRK",
-                    "IPFF",
-                    "SGU",
-                    "ULBI",
-                    "PTIE",
-                    "VMBS",
-                    "WILN",
-                    "DLA",
-                    "FXU",
-                    "PML",
-                    "FTEC",
-                    "MCR",
-                    "PSDV",
-                    "HYGS",
-                    "CVY",
-                    "EUFN",
-                    "TRIV",
-                    "PRTO",
-                    "LITB",
-                    "NATR",
-                    "JSM",
-                    "NWY",
-                    "OESX",
-                    "SCHR",
-                    "HALL",
-                    "CRME",
-                    "FFKT",
-                    "IDI",
-                    "NMZ",
-                    "WMIH",
-                    "DNI",
-                    "EGLE",
-                    "ATAX",
-                    "LMIA",
-                    "APHB",
-                    "AMSC",
-                    "BKEP",
-                    "IFV",
-                    "GARS",
-                    "KTCC",
-                    "NVCN",
-                    "SYMX",
-                    "DHIL",
-                    "KYTH",
-                    "CAPN",
-                    "PPLT",
-                    "GPM",
-                    "BUI",
-                    "HEQ",
-                    "ACSF",
-                    "COB",
-                    "CTQ",
-                    "RETL",
-                    "INVE",
-                    "ZU",
-                    "IRG",
-                    "TNH",
-                    "CRDC",
-                    "GOF",
-                    "PBJ",
-                    "DLS",
-                    "MSL",
-                    "BNO",
-                    "WTBA",
-                    "ALX",
-                    "NRX",
-                    "PRAN",
-                    "IBIO",
-                    "VRP",
-                    "PCF",
-                    "KRNT",
-                    "IUSV",
-                    "AMNB",
-                    "HAP",
-                    "BSRR",
-                    "PSTI",
-                    "IEUR",
-                    "VISI",
-                    "CSD",
-                    "HHY",
-                    "VCV",
-                    "GLMD",
-                    "EDI",
-                    "GDO",
-                    "MFM",
-                    "RLOC",
-                    "SSH",
-                    "VWOB",
-                    "VOC",
-                    "CIO",
-                    "CZNC",
-                    "VNRBP",
-                    "HPS",
-                    "CCNE",
-                    "HYB",
-                    "PRSN",
-                    "WPXP",
-                    "FTF",
-                    "USCI",
-                    "MBLX",
-                    "NID",
-                    "EDIV",
-                    "RWLK",
-                    "FFA",
-                    "FDIS",
-                    "DON",
-                    "ROSG",
-                    "MTLS",
-                    "WHZ",
-                    "TRMR",
-                    "INUV",
-                    "CMBS",
-                    "NMO",
-                    "JDD",
-                    "TPS",
-                    "GASS",
-                    "JLS",
-                    "FCT",
-                    "FOF",
-                    "LEO",
-                    "PWOD",
-                    "TGB",
-                    "JHP",
-                    "LEMB",
-                    "KWEB",
-                    "CRT",
-                    "SPXX",
-                    "SGG",
-                    "DRD",
-                    "FXZ",
-                    "FDL",
-                    "NGE",
-                    "BWEN",
-                    "TRCH",
-                    "NVGN",
-                    "DGICA",
-                    "NSPR",
-                    "CJJD",
-                    "NEA",
-                    "MIXT",
-                    "NQU",
-                    "PFX",
-                    "LGCYO",
-                    "LBMH",
-                    "NRIM",
-                    "HSEA",
-                    "ETO",
-                    "UST",
-                    "MEN",
-                    "IAE",
-                    "ASG",
-                    "WTFCM",
-                    "WPT",
-                    "GOVT",
-                    "BSF",
-                    "SNLN",
-                    "DMB",
-                    "IIM",
-                    "MMU",
-                    "NWLI",
-                    "GTU",
-                    "TACT",
-                    "MLPA",
-                    "DRV",
-                    "TGLS",
-                    "PCEF",
-                    "VTWG",
-                    "KIO",
-                    "USL",
-                    "PXH",
-                    "IAI",
-                    "CBAK",
-                    "VGI",
-                    "AIXG",
-                    "ORPN",
-                    "CVSL",
-                    "BOI",
-                    "FNLC",
-                    "MYN",
-                    "TCPI",
-                    "PHMD",
-                    "GSOL",
-                    "ITM",
-                    "NBB",
-                    "EEMA",
-                    "MHY",
-                    "CYCC",
-                    "IPAC",
-                    "PAR",
-                    "AP",
-                    "RYE",
-                    "BZC",
-                    "SCHP",
-                    "PFMT",
-                    "MGIC",
-                    "ATOS",
-                    "EVN",
-                    "PUB",
-                    "EGIF",
-                    "HEOP",
-                    "AVV",
-                    "MHN",
-                    "MDWD",
-                    "ESCR",
-                    "RDCM",
-                    "XTN",
-                    "PALL",
-                    "AVL",
-                    "ATLO",
-                    "PFXF",
-                    "AFB",
-                    "BSJH",
-                    "WEYS",
-                    "EDUC",
-                    "JMP",
-                    "PFL",
-                    "JSD",
-                    "MUJ",
-                    "FEM",
-                    "FVD",
-                    "FNCL",
-                    "NQP",
-                    "GHI",
-                    "NNVC",
-                    "JGV",
-                    "OCLS",
-                    "BSJI",
-                    "JOEZ",
-                    "EXCU",
-                    "HTR",
-                    "RHS",
-                    "AJX",
-                    "NEAR",
-                    "DWAS",
-                    "CDXS",
-                    "IND",
-                    "SPHQ",
-                    "BSCH",
-                    "BFK",
-                    "PEZ",
-                    "HYLS",
-                    "FBGX",
-                    "DES",
-                    "ITIP",
-                    "NVMI",
-                    "OTIV",
-                    "FRI",
-                    "FTC",
-                    "DRWI",
-                    "ZTR",
-                    "PRH",
-                    "PSF",
-                    "GEK",
-                    "HYLD",
-                    "MDLY",
-                    "JTA",
-                    "FNX",
-                    "GEB",
-                    "LMRK",
-                    "MVO",
-                    "AGD",
-                    "APRI",
-                    "ELON",
-                    "PEY",
-                    "BSCI",
-                    "NXJ",
-                    "FPI",
-                    "HAO",
-                    "NXTD",
-                    "DRA",
-                    "RKDA",
-                    "IIF",
-                    "HOML",
-                    "WAVX",
-                    "NSPH",
-                    "JCAP",
-                    "BYM",
-                    "UTMD",
-                    "NCOM",
-                    "EURL",
-                    "GGT",
-                    "BLBD",
-                    "VGLT",
-                    "FHCO",
-                    "YCS",
-                    "EVM",
-                    "CTV",
-                    "INCO",
-                    "FUND",
-                    "FSAM",
-                    "MGF",
-                    "CXDC",
-                    "CRF",
-                    "BAB",
-                    "OCRX",
-                    "PIP",
-                    "SANW",
-                    "SWJ",
-                    "TINY",
-                    "GBAB",
-                    "MQT",
-                    "INBK",
-                    "RJI",
-                    "HPI",
-                    "CCF",
-                    "ARTX",
-                    "IHD",
-                    "NVET",
-                    "FFNW",
-                    "XLG",
-                    "NPT",
-                    "TLO",
-                    "BME",
-                    "GGOV",
-                    "EIRL",
-                    "WFD",
-                    "FGD",
-                    "IPAS",
-                    "FXC",
-                    "FXA",
-                    "MBII",
-                    "TPZ",
-                    "XENE",
-                    "SQNS",
-                    "WIA",
-                    "BGX",
-                    "FNDF",
-                    "HYMB",
-                    "RELL",
-                    "VOOG",
-                    "CYBE",
-                    "IGOV",
-                    "MXF",
-                    "NKA",
-                    "GHYG",
-                    "DXGE",
-                    "PHT",
-                    "TSI",
-                    "SRLN",
-                    "MCA",
-                    "SBBP",
-                    "SNHN",
-                    "RXI",
-                    "MOBI",
-                    "VZA",
-                    "SGL",
-                    "XHS",
-                    "CBLI",
-                    "III",
-                    "SAMG",
-                    "PROV",
-                    "SCJ",
-                    "NSS",
-                    "PSP",
-                    "ALDX",
-                    "BFZ",
-                    "SPP",
-                    "DWM",
-                    "SRSC",
-                    "FNDA",
-                    "NICK",
-                    "HTBX",
-                    "ISG",
-                    "VPCO",
-                    "LODE",
-                    "ABGB",
-                    "AEK",
-                    "NZF",
-                    "DEX",
-                    "NQM",
-                    "BKT",
-                    "RCPI",
-                    "ISF",
-                    "MOAT",
-                    "VTNR",
-                    "FNDX",
-                    "URRE",
-                    "MHF",
-                    "NL",
-                    "SDOG",
-                    "FALC",
-                    "RVLT",
-                    "YMLP",
-                    "ETX",
-                    "PMX",
-                    "CEW",
-                    "FITBI",
-                    "NQS",
-                    "CHMI",
-                    "MDIV",
-                    "BLMT",
-                    "ONTX",
-                    "NKX",
-                    "IGI",
-                    "MYC",
-                    "PJH",
-                    "SIVR",
-                    "ZSAN",
-                    "ROM",
-                    "NXZ",
-                    "MIRN",
-                    "GEH",
-                    "MTEX",
-                    "MIY",
-                    "SUNS",
-                    "NZH",
-                    "FPE",
-                    "NQI",
-                    "IVAC",
-                    "LCM",
-                    "MYOK",
-                    "RFG",
-                    "PRFZ",
-                    "BTA",
-                    "DWCH",
-                    "CMU",
-                    "OGEN",
-                    "CRED",
-                    "SKYY",
-                    "HNW",
-                    "SOHO",
-                    "CNFR",
-                    "GAIA",
-                    "FEP",
-                    "LONG",
-                    "LAND",
-                    "ADAT",
-                    "VPV",
-                    "SZC",
-                    "INZ",
-                    "HH",
-                    "MHI",
-                    "CIBR",
-                    "NEV",
-                    "APPY",
-                    "VCEL",
-                    "SHI",
-                    "SBW",
-                    "EIS",
-                    "EWV",
-                    "HMPR",
-                    "SQBK",
-                    "QABA",
-                    "RING",
-                    "DVYE",
-                    "CTO",
-                    "DFP",
-                    "IDLV",
-                    "DFJ",
-                    "NAK",
-                    "PWV",
-                    "ZN",
-                    "AEH",
-                    "DEJ",
-                    "GTXI",
-                    "EOT",
-                    "UFPT",
-                    "HDGE",
-                    "GRX",
-                    "MIDU",
-                    "MSON",
-                    "AEUA",
-                    "SMSI",
-                    "MOSY",
-                    "ESLT",
-                    "DHG",
-                    "CHSCL",
-                    "JASN",
-                    "PMD",
-                    "NURO",
-                    "NPF",
-                    "PGLC",
-                    "EIV",
-                    "TLH",
-                    "VTN",
-                    "STPZ",
-                    "HILL",
-                    "BSL",
-                    "SPLP",
-                    "IRIX",
-                    "CXE",
-                    "FFTY",
-                    "HPF",
-                    "MHD",
-                    "PFIS",
-                    "MIND",
-                    "ODC",
-                    "SCX",
-                    "KYO",
-                    "STXS",
-                    "ITR",
-                    "MUE",
-                    "IHC",
-                    "NRCIA",
-                    "ACPW",
-                    "MTCN",
-                    "BNY",
-                    "MGC",
-                    "BBX",
-                    "FAM",
-                    "ADEP",
-                    "GWX",
-                    "WSTL",
-                    "QQEW",
-                    "HCAP",
-                    "GIGM",
-                    "RZA",
-                    "ITI",
-                    "EELV",
-                    "BLE",
-                    "BSJJ",
-                    "CHSCM",
-                    "OFLX",
-                    "DGP",
-                    "FLC",
-                    "PNQI",
-                    "NCTY",
-                    "RINF",
-                    "NLST",
-                    "DUKH",
-                    "SJB",
-                    "RIT",
-                    "VISN",
-                    "NOA",
-                    "OFS",
-                    "NAN",
-                    "PSCH",
-                    "ZHNE",
-                    "RICK",
-                    "BVX",
-                    "TENX",
-                    "GSB",
-                    "USLM",
-                    "JAXB",
-                    "SGA",
-                    "SGOL",
-                    "KOLD",
-                    "KIN",
-                    "YOD",
-                    "IPWR",
-                    "IBND",
-                    "MUA",
-                    "VIOO",
-                    "DAC",
-                    "GEQ",
-                    "PULB",
-                    "DTQ",
-                    "FSGI",
-                    "IDSY",
-                    "FTSL",
-                    "DCIX",
-                    "MQY",
-                    "LTBR",
-                    "SMN",
-                    "VBTX",
-                    "EMF",
-                    "JHY",
-                    "DSS",
-                    "INPH",
-                    "IXP",
-                    "DXJS",
-                    "AMZA",
-                    "ONEQ",
-                    "WINA",
-                    "CHIQ",
-                    "PBP",
-                    "QRHC",
-                    "CCV",
-                    "APLP",
-                    "EAC",
-                    "PSJ",
-                    "CTW",
-                    "RYF",
-                    "NFRA",
-                    "PIE",
-                    "CLUB",
-                    "QTNT",
-                    "IYY",
-                    "GLL",
-                    "TTT",
-                    "EQCO",
-                    "IID",
-                    "WLFC",
-                    "TEU",
-                    "LOR",
-                    "MSK",
-                    "MNOV",
-                    "NUW",
-                    "FHY",
-                    "PZA",
-                    "RJD",
-                    "CDRB",
-                    "FNDE",
-                    "DGRW",
-                    "ADMP",
-                    "PCK",
-                    "PFIE",
-                    "ABIO",
-                    "AXU",
-                    "GURE",
-                    "RZG",
-                    "GSJ",
-                    "CGO",
-                    "NBRV",
-                    "AFSD",
-                    "ALFA",
-                    "URTH",
-                    "DMF",
-                    "MSD",
-                    "INVT",
-                    "LGI",
-                    "CETX",
-                    "NPD",
-                    "SLYG",
-                    "PMF",
-                    "CPIX",
-                    "TDIV",
-                    "LOV",
-                    "DHS",
-                    "MUS",
-                    "FMNB",
-                    "BIF",
-                    "LAG",
-                    "CHN",
-                    "IHDG",
-                    "IMI",
-                    "ROIAK",
-                    "MAV",
-                    "ZROZ",
-                    "AIA",
-                    "KMM",
-                    "MGV",
-                    "IPE",
-                    "FTEK",
-                    "EROC",
-                    "EZM",
-                    "CSM",
-                    "DGZ",
-                    "UYM",
-                    "FDT",
-                    "ACST",
-                    "MVT",
-                    "USDU",
-                    "CAAS",
-                    "NVG",
-                    "FUTY",
-                    "HEEM",
-                    "SSRG",
-                    "CBIO",
-                    "REED",
-                    "RADA",
-                    "SMI",
-                    "APTO",
-                    "RIGS",
-                    "WEB",
-                    "TLI",
-                    "INST",
-                    "GDF",
-                    "DOM",
-                    "HBHCL",
-                    "GAINO",
-                    "NVX",
-                    "ATMP",
-                    "SOCB",
-                    "ANTX",
-                    "SRF",
-                    "PCQ",
-                    "HDNG",
-                    "JKE",
-                    "SMT",
-                    "AGZ",
-                    "PMBC",
-                    "KLDX",
-                    "AFST",
-                    "DIV",
-                    "CDNA",
-                    "KXI",
-                    "FSTA",
-                    "ATAI",
-                    "MARA",
-                    "ISP",
-                    "VONG",
-                    "MXI",
-                    "ORM",
-                    "VGSH",
-                    "BRF",
-                    "CCM",
-                    "CET",
-                    "DHRM",
-                    "NTWK",
-                    "BSJG",
-                    "IAF",
-                    "MNDO",
-                    "APT",
-                    "SPYG",
-                    "BSJF",
-                    "CTX",
-                    "GLV",
-                    "SPHB",
-                    "DARA",
-                    "CVV",
-                    "UTSI",
-                    "GUT",
-                    "CHCI",
-                    "GCVRZ",
-                    "PHF",
-                    "CYRN",
-                    "HDSN",
-                    "JPNL",
-                    "UNL",
-                    "JTD",
-                    "WHLR",
-                    "MYJ",
-                    "PRSS",
-                    "QTEC",
-                    "VGIT",
-                    "MIDZ",
-                    "NBH",
-                    "TSLF",
-                    "EMHY",
-                    "MDM",
-                    "NRT",
-                    "FULL",
-                    "KOL",
-                    "GSLC",
-                    "AAU",
-                    "PIZ",
-                    "GGE",
-                    "FREL",
-                    "ILB",
-                    "FRBK",
-                    "FPXI",
-                    "TST",
-                    "MSP",
-                    "PBS",
-                    "COPX",
-                    "PPX",
-                    "FNI",
-                    "BHB",
-                    "PZC",
-                    "IQDF",
-                    "DNO",
-                    "VBF",
-                    "ECT",
-                    "REFR",
-                    "HIIQ",
-                    "EXD",
-                    "ACFC",
-                    "GAINN",
-                    "DTN",
-                    "UZB",
-                    "KINS",
-                    "SYNC",
-                    "FCO",
-                    "GLYC",
-                    "WLRH",
-                    "LMNS",
-                    "CQQQ",
-                    "ADMA",
-                    "CPAC",
-                    "FYX",
-                    "INTX",
-                    "IWY",
-                    "TBIO",
-                    "BSCK",
-                    "LUB",
-                    "IVOO",
-                    "LWC",
-                    "EDGW",
-                    "MPX",
-                    "MGR",
-                    "CTU",
-                    "DTZ",
-                    "HTGM",
-                    "UDN",
-                    "IOO",
-                    "NGHCZ",
-                    "SKIS",
-                    "VONE",
-                    "DCUC",
-                    "MWR",
-                    "TPL",
-                    "FEU",
-                    "BPTH",
-                    "MCBC",
-                    "SVVC",
-                    "GMM",
-                    "SGDM",
-                    "TLOG",
-                    "TACOW",
-                    "GURU",
-                    "JGBD",
-                    "BOTA",
-                    "GMO",
-                    "EMMS",
-                    "CORP",
-                    "WEA",
-                    "TSNU",
-                    "DGRO",
-                    "AOA",
-                    "FRPH",
-                    "OSGB",
-                    "AEB",
-                    "MLPW",
-                    "PLPC",
-                    "NCA",
-                    "PBW",
-                    "IPCI",
-                    "BBRC",
-                    "DOL",
-                    "BKN",
-                    "NXQ",
-                    "RCD",
-                    "ESNC",
-                    "CRDS",
-                    "RGT",
-                    "HPJ",
-                    "CEMI",
-                    "EBTC",
-                    "VIXM",
-                    "CCUR",
-                    "EGAS",
-                    "GLOW",
-                    "QUMU",
-                    "PPA",
-                    "FILL",
-                    "ERH",
-                    "LUNA",
-                    "YECO",
-                    "UGL",
-                    "LOJN",
-                    "CYAD",
-                    "LEU",
-                    "XRA",
-                    "RYU",
-                    "LEI",
-                    "GRO",
-                    "GDL",
-                    "GSIT",
-                    "VLT",
-                    "IMLP",
-                    "TAYD",
-                    "GSF",
-                    "MBVT",
-                    "ULTR",
-                    "EMCB",
-                    "JHI",
-                    "HIFS",
-                    "SYLD",
-                    "CORN",
-                    "RESN",
-                    "LIVE",
-                    "MNP",
-                    "CGW",
-                    "CXH",
-                    "UZA",
-                    "MYF",
-                    "EXFO",
-                    "SOCL",
-                    "BSCJ",
-                    "GSY",
-                    "INP",
-                    "VNRAP",
-                    "NGHCO",
-                    "PWB",
-                    "DBB",
-                    "AOR",
-                    "PHDG",
-                    "EARS",
-                    "BSCG",
-                    "SLYV",
-                    "FNDC",
-                    "SMIN",
-                    "ADXSW",
-                    "PDII",
-                    "SWSH",
-                    "NERV",
-                    "GLDX",
-                    "MLPX",
-                    "HNSN",
-                    "KHI",
-                    "CHSCN",
-                    "GFY",
-                    "CUBA",
-                    "MNA",
-                    "EDEN",
-                    "MBTF",
-                    "SMHD",
-                    "IHY",
-                    "BAF",
-                    "CRD.A",
-                    "ROKA",
-                    "FIDU",
-                    "TRF",
-                    "VTRB",
-                    "FRAK",
-                    "SITO",
-                    "OPTT",
-                    "MARK",
-                    "MWO",
-                    "LIQD",
-                    "NORW",
-                    "SALM",
-                    "MDYG",
-                    "EXI",
-                    "AVNU",
-                    "JVA",
-                    "OHAI",
-                    "SPFF",
-                    "DWTR",
-                    "TI.A",
-                    "AFW",
-                    "MPA",
-                    "MTT",
-                    "DRAM",
-                    "BANX",
-                    "MUNI",
-                    "AXTI",
-                    "USEG",
-                    "OIIM",
-                    "PCYO",
-                    "SIFY",
-                    "CNTF",
-                    "FCVA",
-                    "AGNCB",
-                    "IAK",
-                    "LIQT",
-                    "ARGS",
-                    "TDA",
-                    "HBCP",
-                    "AFGE",
-                    "EMFM",
-                    "PRTS",
-                    "EWO",
-                    "WEAT",
-                    "FRN",
-                    "PEJ",
-                    "SPPP",
-                    "SWU",
-                    "BNTC",
-                    "KNOW",
-                    "DBGR",
-                    "IXG",
-                    "KBWD",
-                    "COMT",
-                    "FBG",
-                    "AE",
-                    "SUB",
-                    "JHX",
-                    "BCBP",
-                    "VONV",
-                    "ORRF",
-                    "KFH",
-                    "MWG",
-                    "GSV",
-                    "AOM",
-                    "FJP",
-                    "PNI",
-                    "GSH",
-                    "NTIP",
-                    "ZAIS",
-                    "PFD",
-                    "ASRVP",
-                    "ENX",
-                    "CIFC",
-                    "FLTB",
-                    "CAMT",
-                    "NUM",
-                    "SIFI",
-                    "DTD",
-                    "DUC",
-                    "NBD",
-                    "CNCR",
-                    "MUH",
-                    "SONA",
-                    "BKF",
-                    "DBKO",
-                    "APF",
-                    "RNO",
-                    "MZF",
-                    "MFD",
-                    "BBK",
-                    "RCON",
-                    "NHLD",
-                    "STDY",
-                    "SKLN",
-                    "EFZ",
-                    "PGJ",
-                    "NXP",
-                    "AWRE",
-                    "VIOG",
-                    "BKK",
-                    "WOOD",
-                    "AMRK",
-                    "ARCX",
-                    "UWN",
-                    "KIQ",
-                    "CFD",
-                    "JOB",
-                    "FGB",
-                    "DAKP",
-                    "VRML",
-                    "AUMN",
-                    "VYGR",
-                    "JJG",
-                    "FMD",
-                    "ARCW",
-                    "CETC",
-                    "GIGA",
-                    "PDN",
-                    "TLTD",
-                    "OSM",
-                    "EFF",
-                    "PCMI",
-                    "LGCYP",
-                    "ETAK",
-                    "ITE",
-                    "FNDB",
-                    "MCI",
-                    "SNFCA",
-                    "JPW",
-                    "JXI",
-                    "ASM",
-                    "IKGH",
-                    "SWZ",
-                    "RIF",
-                    "DSUM",
-                    "SKBI",
-                    "DVHL",
-                    "FAV",
-                    "SGZA",
-                    "QUANTCONNECT-PRN",
-                    "STS",
-                    "ASTC",
-                    "AED",
-                    "BSCM",
-                    "DTH",
-                    "AMTX",
-                    "IGN",
-                    "IFEU",
-                    "NYMTO",
-                    "RMCF",
-                    "GDEN",
-                    "CEV",
-                    "NCQ",
-                    "ASFI",
-                    "CRWS",
-                    "FIGY",
-                    "CTF",
-                    "SMEZ",
-                    "DBP",
-                    "MTK",
-                    "DBE",
-                    "MOLG",
-                    "RWL",
-                    "ADK",
-                    "BCRH",
-                    "CVOL",
-                    "NTRSP",
-                    "DOO",
-                    "RJA",
-                    "ENG",
-                    "NMY",
-                    "BWFG",
-                    "CHSCO",
-                    "SOR",
-                    "HDLV",
-                    "PRGN",
-                    "SEED",
-                    "VIRC",
-                    "JAGX",
-                    "EUSC",
-                    "JETS",
-                    "GNVC",
-                    "CNET",
-                    "ENZL",
-                    "DTYS",
-                    "AFSS",
-                    "AFK",
-                    "CHIX",
-                    "FKU",
-                    "XPLR",
-                    "WDTI",
-                    "CLRO",
-                    "EML",
-                    "ISTB",
-                    "PFO",
-                    "MELA",
-                    "CRHM",
-                    "EAE",
-                    "ATE",
-                    "HIHO",
-                    "TTF",
-                    "UNTY",
-                    "GGZ",
-                    "PTLC",
-                    "EAB",
-                    "IYLD",
-                    "FAB",
-                    "NIQ",
-                    "LENS",
-                    "VBIV",
-                    "GGM",
-                    "MGT",
-                    "KSM",
-                    "TDJ",
-                    "GV",
-                    "ENSV",
-                    "SLX",
-                    "TLTE",
-                    "UPLD",
-                    "AOK",
-                    "KOOL",
-                    "REE",
-                    "QQQE",
-                    "CHSCP",
-                    "BIZD",
-                    "TTFS",
-                    "GCC",
-                    "CXA",
-                    "PTNR",
-                    "SEA",
-                    "TISA",
-                    "NUO",
-                    "GFN",
-                    "TCBIP",
-                    "HSON",
-                    "ECF",
-                    "CGNT",
-                    "PICB",
-                    "QQXT",
-                    "NKY",
-                    "EAA",
-                    "GDXS",
-                    "CGA",
-                    "RLOG",
-                    "BXC",
-                    "IOTS",
-                    "JKG",
-                    "DGLD",
-                    "RCMT",
-                    "KF",
-                    "RIVR",
-                    "KGJI",
-                    "FDD",
-                    "QLTA",
-                    "RWJ",
-                    "RVSB",
-                    "CGG",
-                    "FTGC",
-                    "CBSHP",
-                    "PXI",
-                    "GLRI",
-                    "MSF",
-                    "FSZ",
-                    "MTS",
-                    "CSI",
-                    "XAR",
-                    "DCUB",
-                    "SORL",
-                    "ZMLP",
-                    "AGIIL",
-                    "FYC",
-                    "KFS",
-                    "GBF",
-                    "BLFS",
-                    "AKER",
-                    "JOF",
-                    "THM",
-                    "RWK",
-                    "CNBKA",
-                    "FT",
-                    "PNF",
-                    "AKP",
-                    "MMAC",
-                    "IVOG",
-                    "BPK",
-                    "LPTH",
-                    "ENFC",
-                    "CDC",
-                    "AINC",
-                    "CAPR",
-                    "FXB",
-                    "UBT",
-                    "PCM",
-                    "JKD",
-                    "UUU",
-                    "CSUN",
-                    "GROW",
-                    "AKTX",
-                    "HYHG",
-                    "ACTS",
-                    "AFTY",
-                    "PACEU",
-                    "PIO",
-                    "EGAN",
-                    "NNY",
-                    "ULE",
-                    "AAME",
-                    "REMX",
-                    "RRM",
-                    "FYLD",
-                    "TANP",
-                    "EMITF",
-                    "DZK",
-                    "WSTG",
-                    "FHK",
-                    "DCUA",
-                    "EAGLU",
-                    "MTR",
-                    "CIF",
-                    "ICLN",
-                    "UMDD",
-                    "JHS",
-                    "FSI",
-                    "FXEU",
-                    "ALXA",
-                    "VNRX",
-                    "FMAT",
-                    "AFC",
-                    "JKF",
-                    "ELJ",
-                    "IQDE",
-                    "PSCT",
-                    "CYRX",
-                    "FDI",
-                    "SDYL",
-                    "GRH",
-                    "FTCS",
-                    "MLPC",
-                    "AIY",
-                    "SEF",
-                    "NNC",
-                    "EMAN",
-                    "PST",
-                    "ISRL",
-                    "STRM",
-                    "BSCL",
-                    "FNJN",
-                    "ENVI",
-                    "RSXJ",
-                    "DIVY",
-                    "UAMY",
-                    "EES",
-                    "FGM",
-                    "TRNX",
-                    "CYHHZ",
-                    "DSI",
-                    "OUSA",
-                    "HGSH",
-                    "SUMR",
-                    "BSCF",
-                    "GRP.U",
-                    "EWZS",
-                    "LOAN",
-                    "AUPH",
-                    "SSN",
-                    "ECC",
-                    "IDN",
-                    "MHNC",
-                    "FIEG",
-                    "REK",
-                    "CREG",
-                    "LIT",
-                    "CUT",
-                    "FNWB",
-                    "FCOM",
-                    "KCE",
-                    "AIB",
-                    "NTC",
-                    "AGNCP",
-                    "JJC",
-                    "FNY",
-                    "ROYL",
-                    "TTNP",
-                    "CMF",
-                    "OPGN",
-                    "LRAD",
-                    "FEEU",
-                    "DGL",
-                    "DPK",
-                    "IESC",
-                    "SDPI",
-                    "PXE",
-                    "BABS",
-                    "GF",
-                    "GLTR",
-                    "EPAX",
-                    "BAGR",
-                    "IPDN",
-                    "DBV",
-                    "RMGN",
-                    "SAR",
-                    "GSI",
-                    "BOCH",
-                    "LVL",
-                    "VGZ",
-                    "FTSM",
-                    "LINDW",
-                    "RFTA",
-                    "PBIP",
-                    "JCS",
-                    "YUMA",
-                    "IDOG",
-                    "DJCO",
-                    "STIP",
-                    "VLUE",
-                    "BIK",
-                    "CASM",
-                    "FIEU",
-                    "PULM",
-                    "FES",
-                    "FXF",
-                    "AGFSW",
-                    "WILC",
-                    "BSE",
-                    "NIB",
-                    "AAAP",
-                    "DDF",
-                    "BBC",
-                    "CNYD",
-                    "GILT",
-                    "FLL",
-                    "BSPM",
-                    "PSAU",
-                    "OXLCO",
-                    "FEO",
-                    "FXENP",
-                    "PFBI",
-                    "LBJ",
-                    "MORT",
-                    "NBW",
-                    "OSHC",
-                    "BPMX",
-                    "MDYV",
-                    "IRY",
-                    "JP",
-                    "RZV",
-                    "ICB",
-                    "SLI",
-                    "OLBK",
-                    "RLJE",
-                    "HSGX",
-                    "FAC",
-                    "SPE",
-                    "JST",
-                    "EGF",
-                    "QKLS",
-                    "KLD",
-                    "PAI",
-                    "SXCL",
-                    "JKL",
-                    "JPMV",
-                    "BNJ",
-                    "MPET",
-                    "BWZ",
-                    "KBWP",
-                    "FIW",
-                    "SLY",
-                    "DVCR",
-                    "MSLI",
-                    "VOOV",
-                    "ISNS",
-                    "RTTR",
-                    "XSLV",
-                    "NVFY",
-                    "PFM",
-                    "DEF",
-                    "PGAL",
-                    "GCV",
-                    "QDF",
-                    "TDE",
-                    "JMT",
-                    "EVY",
-                    "ROBO",
-                    "AMRB",
-                    "ACUR",
-                    "AETI",
-                    "TILT",
-                    "TLF",
-                    "EIA",
-                    "FSNN",
-                    "PESI",
-                    "RGI",
-                    "ARIS",
-                    "NXR",
-                    "OTEL",
-                    "BFY",
-                    "FCAM",
-                    "BDCS",
-                    "MSCA",
-                    "GLU",
-                    "KBWY",
-                    "SLVP",
-                    "EFNL",
-                    "GCH",
-                    "ENOR",
-                    "XT",
-                    "HNNA",
-                    "GXF",
-                    "TKF",
-                    "HUSA",
-                    "NMI",
-                    "MYOS",
-                    "INFU",
-                    "DAG",
-                    "CH",
-                    "GSP",
-                    "CEE",
-                    "ARGT",
-                    "HYXU",
-                    "CRVP",
-                    "CBS.A",
-                    "EET",
-                    "IF",
-                    "BGCA",
-                    "CMCT",
-                    "BBP",
-                    "PGZ",
-                    "PSCE",
-                    "SIM",
-                    "GDXX",
-                    "SMDD",
-                    "SGMA",
-                    "SKY",
-                    "ESTE",
-                    "NMT",
-                    "XPL",
-                    "MFV",
-                    "MFT",
-                    "SAQ",
-                    "KCNY",
-                    "SIF",
-                    "SST",
-                    "CPHR",
-                    "CYB",
-                    "EBND",
-                    "HBP",
-                    "DVYL",
-                    "THGA",
-                    "NPV",
-                    "TAI",
-                    "IDSA",
-                    "SLRA",
-                    "YYY",
-                    "TOK",
-                    "NTEC",
-                    "NFLT",
-                    "NTX",
-                    "MBG",
-                    "MPV",
-                    "CFRXZ",
-                    "TAOM",
-                    "MGH",
-                    "SLVO",
-                    "WDIV",
-                    "BORN",
-                    "VALU",
-                    "FYT",
-                    "ANGL",
-                    "MGN",
-                    "DBEZ",
-                    "AQMS",
-                    "VTEB",
-                    "EZT",
-                    "ISSC",
-                    "VMM",
-                    "FLRN",
-                    "ITRN",
-                    "GBLIZ",
-                    "PSI",
-                    "BSJK",
-                    "ESSX",
-                    "TVE",
-                    "EGPT",
-                    "ADGE",
-                    "PLW",
-                    "ECTE",
-                    "VKTX",
-                    "JFC",
-                    "IPKW",
-                    "ICCC",
-                    "BZF",
-                    "MYY",
-                    "PZG",
-                    "QCLN",
-                    "CFFI",
-                    "KMPA",
-                    "BTN",
-                    "TANN",
-                    "INOD",
-                    "COW",
-                    "GQRE",
-                    "INTT",
-                    "VXDN",
-                    "ASBI",
-                    "GBIM",
-                    "KTP",
-                    "NBO",
-                    "SLMAP",
-                    "SMB",
-                    "BCV",
-                    "CFO",
-                    "GAL",
-                    "SBI",
-                    "LAQ",
-                    "LRE",
-                    "MFRI",
-                    "XHE",
-                    "EZJ",
-                    "DTF",
-                    "ARU",
-                    "EMXX",
-                    "SGYPW",
-                    "TANO",
-                    "CHNR",
-                    "SRT",
-                    "RWC",
-                    "PTMC",
-                    "SPRO",
-                    "VSR",
-                    "MNRK",
-                    "PSCD",
-                    "WOWO",
-                    "AMOV",
-                    "ILTB",
-                    "ASRV",
-                    "DXJH",
-                    "GRIF",
-                    "BSD",
-                    "GEF.B",
-                    "MAGS",
-                    "ESBA",
-                    "NAII",
-                    "EFU",
-                    "ESSA",
-                    "CALI",
-                    "ONFC",
-                    "ESES",
-                    "WBII",
-                    "XMLV",
-                    "FAD",
-                    "ROOF",
-                    "GVAL",
-                    "AMSGP",
-                    "WTT",
-                    "CVU",
-                    "TVC",
-                    "SGF",
-                    "IBDH",
-                    "KBWR",
-                    "WGA",
-                    "WLRHU",
-                    "DGRE",
-                    "TDI",
-                    "JPGE",
-                    "NAUH",
-                    "PSCF",
-                    "ANCB",
-                    "FTLS",
-                    "LLEX",
-                    "LBF",
-                    "CHII",
-                    "ADRE",
-                    "PXJ",
-                    "SHBI",
-                    "FPT",
-                    "ATNY",
-                    "MAMS",
-                    "RRF",
-                    "TGC",
-                    "NYMTP",
-                    "VTWV",
-                    "ENRJ",
-                    "JEQ",
-                    "ALTS",
-                    "ASEA",
-                    "BFOR",
-                    "DSXN",
-                    "UBFO",
-                    "GYRO",
-                    "EEB",
-                    "GFNSL",
-                    "STLY",
-                    "FPA",
-                    "BBF",
-                    "NXC",
-                    "RAVI",
-                    "GUR",
-                    "MDVX",
-                    "EMZ",
-                    "CBMX",
-                    "SGOC",
-                    "CAW",
-                    "DYSL",
-                    "MOCO",
-                    "DDT",
-                    "SIJ",
-                    "EVJ",
-                    "PRXI",
-                    "DRR",
-                    "DHVW",
-                    "CARV",
-                    "SWH",
-                    "VIOV",
-                    "GENC",
-                    "BLIN",
-                    "ENY",
-                    "HEWI",
-                    "IEUS",
-                    "ISHG",
-                    "IGHG",
-                    "MTBC",
-                    "FPP",
-                    "VXUP",
-                    "MINC",
-                    "WF",
-                    "MDD",
-                    "FAN",
-                    "PYDS",
-                    "GII",
-                    "JKI",
-                    "RFIL",
-                    "SPYV",
-                    "BJK",
-                    "MLN",
-                    "EMIF",
-                    "ALTV",
-                    "BDCV",
-                    "JKJ",
-                    "QVAL",
-                    "DRW",
-                    "PBIB",
-                    "FMN",
-                    "CCA",
-                    "JYNT",
-                    "NTN",
-                    "BTAL",
-                    "VQT",
-                    "TWMC",
-                    "PKB",
-                    "KOSS",
-                    "PSCU",
-                    "TDTF",
-                    "SGDJ",
-                    "USD",
-                    "FLGE",
-                    "XRLV",
-                    "BYLK",
-                    "RYJ",
-                    "IUSB",
-                    "OXLCN",
-                    "GRES",
-                    "MEIL",
-                    "RIBT",
-                    "PED",
-                    "RLY",
-                    "CTHR",
-                    "PNTR",
-                    "RFV",
-                    "NAZ",
-                    "GVP",
-                    "KST",
-                    "LCNB",
-                    "IEIL",
-                    "NCB",
-                    "SBFG",
-                    "SOFO",
-                    "FLTR",
-                    "EWGS",
-                    "HVPW",
-                    "PKBK",
-                    "BKSC",
-                    "NIM",
-                    "ARKR",
-                    "NYF",
-                    "TAS",
-                    "EVBS",
-                    "PTNQ",
-                    "ISRA",
-                    "BBDO",
-                    "FWDI",
-                    "KUTV",
-                    "NYV",
-                    "TUZ",
-                    "USBI",
-                    "DDE",
-                    "CELGZ",
-                    "QYLD",
-                    "XMPT",
-                    "AFA",
-                    "YXI",
-                    "XELB",
-                    "WVVI",
-                    "ACU",
-                    "JPIN",
-                    "ITUS",
-                    "NFO",
-                    "MHNB",
-                    "COOL",
-                    "FBND",
-                    "RTM",
-                    "CU",
-                    "PWZ",
-                    "TIK",
-                    "PARN",
-                    "SMTP",
-                    "WLRHW",
-                    "ELA",
-                    "EUMV",
-                    "WVE",
-                    "NKG",
-                    "GAZ",
-                    "FFR",
-                    "IWX",
-                    "ECCA",
-                    "HTUS",
-                    "PIH",
-                    "VCF",
-                    "IBDL",
-                    "MLSS",
-                    "IBLN",
-                    "BAL",
-                    "CRAK",
-                    "NXN",
-                    "SMMF",
-                    "IEC",
-                    "PBD",
-                    "FNK",
-                    "ABRN",
-                    "EWRM",
-                    "IRL",
-                    "CLRX",
-                    "TOF",
-                    "BVSN",
-                    "MVCB",
-                    "EQL",
-                    "OPHC",
-                    "DSWL",
-                    "CVCY",
-                    "DEW",
-                    "VTHR",
-                    "REW",
-                    "AMCO",
-                    "PLTM",
-                    "GUA",
-                    "LTPZ",
-                    "JMM",
-                    "IVOV",
-                    "LDF",
-                    "MOKO",
-                    "DVP",
-                    "INTG",
-                    "MXE",
-                    "CARO",
-                    "HUB.A",
-                    "DGRS",
-                    "GULTU",
-                    "APB",
-                    "ESXB",
-                    "NTZ",
-                    "RDIV",
-                    "IPW",
-                    "ACIM",
-                    "ECNS",
-                    "REET",
-                    "PFH",
-                    "TUTT",
-                    "AEHR",
-                    "EEA",
-                    "ICOL",
-                    "SILJ",
-                    "EGI",
-                    "VFL",
-                    "PYN",
-                    "ELMD",
-                    "HKTV",
-                    "IBDK",
-                    "XTL",
-                    "EVAR",
-                    "FUNC",
-                    "IBKCP",
-                    "SFST",
-                    "SLTB",
-                    "SBLKL",
-                    "VIDI",
-                    "AIRT",
-                    "CZA",
-                    "KORU",
-                    "MZZ",
-                    "TSBK",
-                    "DIM",
-                    "DXPS",
-                    "XBKS",
-                    "UCI",
-                    "IGLD",
-                    "QLTB",
-                    "SCYX",
-                    "PVBC",
-                    "ONP",
-                    "WPS",
-                    "IPAY",
-                    "BIE",
-                    "RBPAA",
-                    "EVI",
-                    "BVXV",
-                    "ELU",
-                    "RGCO",
-                    "RVP",
-                    "VUSE",
-                    "AIRI",
-                    "IBDM",
-                    "DVD",
-                    "PLND",
-                    "GLDI",
-                    "BJZ",
-                    "MBRG",
-                    "SOYB",
-                    "KMDA",
-                    "ISIG",
-                    "UG",
-                    "GNMA",
-                    "MFSF",
-                    "MLPY",
-                    "JKK",
-                    "MGCD",
-                    "FSBK",
-                    "SHLDW",
-                    "CJNK",
-                    "DGAS",
-                    "KEF",
-                    "MNE",
-                    "PGHY",
-                    "TLR",
-                    "MFNC",
-                    "DVYA",
-                    "KBWC",
-                    "FSCE",
-                    "JPP",
-                    "ABE",
-                    "WRN",
-                    "GULF",
-                    "NJV",
-                    "SSY",
-                    "UBR",
-                    "IGC",
-                    "AMBCW",
-                    "UFAB",
-                    "FCCY",
-                    "FEMS",
-                    "GPRK",
-                    "PFIN",
-                    "RDVY",
-                    "CLNT",
-                    "FDM",
-                    "CARZ",
-                    "IBDJ",
-                    "BLJ",
-                    "AHPI",
-                    "SHIP",
-                    "VII",
-                    "WBIF",
-                    "WSCI",
-                    "IELG",
-                    "NMS",
-                    "INXX",
-                    "BSCN",
-                    "PSCC",
-                    "CCSC",
-                    "SSWN",
-                    "KONE",
-                    "ELB",
-                    "ARCI",
-                    "BSCO",
-                    "YGRO",
-                    "BQH",
-                    "GEX",
-                    "XSW",
-                    "YMLI",
-                    "AWX",
-                    "PRZM",
-                    "FSCFL",
-                    "FMY",
-                    "ELTK",
-                    "DDG",
-                    "OUNZ",
-                    "PEBK",
-                    "WBIH",
-                    "SHYD",
-                    "CLRB",
-                    "SBND",
-                    "JTPY",
-                    "OASM",
-                    "REDF",
-                    "WMGIZ",
-                    "WBIG",
-                    "OPOF",
-                    "KWT",
-                    "GSC",
-                    "RALS",
-                    "LSBG",
-                    "BELFA",
-                    "JPXN",
-                    "PRPH",
-                    "QDEF",
-                    "UAE",
-                    "ABAC",
-                    "FSBW",
-                    "CBNK",
-                    "DNL",
-                    "HBMD",
-                    "WBIC",
-                    "JKH",
-                    "GLDC",
-                    "LRET",
-                    "SPUU",
-                    "PWC",
-                    "EVGN",
-                    "SBRAP",
-                    "ZBK",
-                    "BYBK",
-                    "RBL",
-                    "BLH",
-                    "YAO",
-                    "DXJF",
-                    "EQS",
-                    "GAINP",
-                    "ADRD",
-                    "UCC",
-                    "DGSE",
-                    "TPI",
-                    "WBID",
-                    "CIZ",
-                    "BAMM",
-                    "EPS",
-                    "TWOK",
-                    "TGEN",
-                    "CROC",
-                    "DAIO",
-                    "UPW",
-                    "DWAQ",
-                    "KBA",
-                    "GML",
-                    "EUSA",
-                    "DBAW",
-                    "WBIB",
-                    "GYB",
-                    "BFO",
-                    "CHIM",
-                    "IST",
-                    "QQQC",
-                    "SAEX",
-                    "IBMH",
-                    "ERS",
-                    "EMQ",
-                    "DXR",
-                    "FRD",
-                    "TCCB",
-                    "RFT",
-                    "IPO",
-                    "TPVZ",
-                    "WBIE",
-                    "PYZ",
-                    "GYEN",
-                    "EEI",
-                    "PSCI",
-                    "PDEX",
-                    "SPU",
-                    "WBIA",
-                    "CSF",
-                    "EMQQ",
-                    "EWRS",
-                    "CERCU",
-                    "LEDS",
-                    "EMJ",
-                    "IBMG",
-                    "DXJR",
-                    "LIME",
-                    "IWL",
-                    "CPHI",
-                    "AEY",
-                    "MOM",
-                    "AIRR",
-                    "CHMG",
-                    "CAFE",
-                    "CHOP",
-                    "NYH",
-                    "EBMT",
-                    "EQAL",
-                    "IBMI",
-                    "SBNA",
-                    "CIVB",
-                    "PAGG",
-                    "SMTX",
-                    "PXLG",
-                    "TWN",
-                    "INNL",
-                    "SZO",
-                    "PFI",
-                    "INKM",
-                    "IBDN",
-                    "EYEGW",
-                    "HFBC",
-                    "BGSF",
-                    "TYO",
-                    "NLR",
-                    "CRBN",
-                    "FMB",
-                    "FEIM",
-                    "HFFC",
-                    "SPAN",
-                    "IBMF",
-                    "BKEPP",
-                    "PXSV",
-                    "PZT",
-                    "ARDM",
-                    "EFUT",
-                    "GEUR",
-                    "PXMG",
-                    "CHEV",
-                    "SLQD",
-                    "UBOH",
-                    "BDR",
-                    "ARTW",
-                    "SEV",
-                    "CCCL",
-                    "AUSE",
-                    "GRI",
-                    "SVBI",
-                    "MLPJ",
-                    "NVSL",
-                    "ISL",
-                    "BTH",
-                    "QADB",
-                    "BUNT",
-                    "DLBS",
-                    "CLBH",
-                    "VEGI",
-                    "YLD",
-                    "CCCR",
-                    "ACNB",
-                    "HOVNP",
-                    "VBND",
-                    "SSFN",
-                    "PVTD",
-                    "PFIG",
-                    "CYRXW",
-                    "LTRX",
-                    "FTY",
-                    "MSN",
-                    "TOPS",
-                    "INTLL",
-                    "CRBQ",
-                    "ESMC",
-                    "HFXE",
-                    "SMDV",
-                    "FBMS",
-                    "MHNA",
-                    "OVBC",
-                    "ISTR",
-                    "MLP",
-                    "KFFB",
-                    "KFI",
-                    "DAX",
-                    "EFAD",
-                    "UNB",
-                    "DRWIW",
-                    "PTM",
-                    "BSJL",
-                    "HIPS",
-                    "SBNB",
-                    "HSPX",
-                    "BSI",
-                    "DBBR",
-                    "TMUSP",
-                    "TAO",
-                    "PMR",
-                    "MFO",
-                    "YCB",
-                    "STKS",
-                    "CSPI",
-                    "THRK",
-                    "LDUR",
-                    "IBDF",
-                    "ATV",
-                    "SIEB",
-                    "SMLR",
-                    "CHEK",
-                    "IBME",
-                    "RXD",
-                    "WMCR",
-                    "EUFX",
-                    "TOLZ",
-                    "BRT",
-                    "CWAY",
-                    "HMNF",
-                    "CRTN",
-                    "NBN",
-                    "LOOK",
-                    "AGND",
-                    "CYAN",
-                    "GNAT",
-                    "STRS",
-                    "VCO",
-                    "EWUS",
-                    "NWBOW",
-                    "GRN",
-                    "CSLS",
-                    "MLVF",
-                    "SFN",
-                    "HFXI",
-                    "LMBS",
-                    "EVK",
-                    "CBAN",
-                    "RJN",
-                    "CBON",
-                    "MHH",
-                    "BOSC",
-                    "PSR",
-                    "EIP",
-                    "PERM",
-                    "HTGX",
-                    "EFM",
-                    "GRR",
-                    "PXR",
-                    "PNTA",
-                    "BSJM",
-                    "OXBR",
-                    "HGI",
-                    "MARPS",
-                    "SLCT",
-                    "PUI",
-                    "GEM",
-                    "FVL",
-                    "SKYS",
-                    "CBND",
-                    "WBIL",
-                    "PANL",
-                    "PLBC",
-                    "HEWP",
-                    "TIPZ",
-                    "TCCA",
-                    "SAL",
-                    "FFNM",
-                    "ENFR",
-                    "CPAH",
-                    "REGL",
-                    "SLMBP",
-                    "CANE",
-                    "CPSH",
-                    "MHE",
-                    "UPV",
-                    "PEX",
-                    "PWX",
-                    "SSG",
-                    "DI",
-                    "PME",
-                    "BRAZ",
-                    "BPOPN",
-                    "TBX",
-                    "RELV",
-                    "IQDY",
-                    "IPU",
-                    "SBSA",
-                    "ITIC",
-                    "LPSB",
-                    "RWW",
-                    "PUW",
-                    "HYZD",
-                    "HILO",
-                    "EWRI",
-                    "YDIV",
-                    "AUBN",
-                    "PXMV",
-                    "IVAL",
-                    "PYS",
-                    "GAA",
-                    "CEMB",
-                    "ALN",
-                    "CLA",
-                    "AXJL",
-                    "CZWI",
-                    "BRAQ",
-                    "ULST",
-                    "TCBIL",
-                    "PW",
-                    "THST",
-                    "SMLV",
-                    "EIO",
-                    "LALT",
-                    "UBP",
-                    "OCLSW",
-                    "JMPB",
-                    "PZD",
-                    "SRET",
-                    "SMBC",
-                    "EMDD",
-                    "MCV",
-                    "BVA",
-                    "DBS",
-                    "OSBCP",
-                    "IBDQ",
-                    "JJA",
-                    "IFMI",
-                    "ALD",
-                    "IBDP",
-                    "DYNT",
-                    "NWFL",
-                    "VAMO",
-                    "ROIA",
-                    "SMMT",
-                    "ENJ",
-                    "FINU",
-                    "MTSL",
-                    "ROIQW",
-                    "KRU",
-                    "AUNZ",
-                    "IBDO",
-                    "DJCI",
-                    "SINO",
-                    "JHML",
-                    "CVLY",
-                    "ONEK",
-                    "DLHC",
-                    "IPS",
-                    "OPXAW",
-                    "ERB",
-                    "TOR",
-                    "JBN",
-                    "PXQ",
-                    "MAB",
-                    "PBHC",
-                    "OGCP",
-                    "TRNS",
-                    "IIJI",
-                    "ADRA",
-                    "OCC",
-                    "UXI",
-                    "PVTBP",
-                    "MJCO",
-                    "EDOG",
-                    "ACY",
-                    "FCTY",
-                    "FRBA",
-                    "UBCP",
-                    "CIZN",
-                    "EZY",
-                    "DXKW",
-                    "PPHMP",
-                    "TUTI",
-                    "EWEM",
-                    "PATI",
-                    "ARKW",
-                    "BBGI",
-                    "HDG",
-                    "LARK",
-                    "FLM",
-                    "NRCIB",
-                    "PDBC",
-                    "UCBA",
-                    "QLTC",
-                    "NOM",
-                    "XLFS",
-                    "EMCD",
-                    "PXSG",
-                    "CBFV",
-                    "PFK",
-                    "EGT",
-                    "AGZD",
-                    "GZT",
-                    "OVLY",
-                    "HTGY",
-                    "CPHC",
-                    "BWL.A",
-                    "POPE",
-                    "MCOX",
-                    "AKO.A",
-                    "CTIB",
-                    "WFBI",
-                    "BOCA",
-                    "YLCO",
-                    "PVI",
-                    "TRCB",
-                    "SBCP",
-                    "AIC",
-                    "BASI",
-                    "MZA",
-                    "GMOM",
-                    "MEAR",
-                    "KWN",
-                    "RSCO",
-                    "VSPY",
-                    "IBDC",
-                    "PERF",
-                    "EVO",
-                    "HYGH",
-                    "QAUS",
-                    "VALX",
-                    "AXR",
-                    "CWBC",
-                    "QMN",
-                    "SBBX",
-                    "ALOT",
-                    "OLO",
-                    "SYG",
-                    "JASNW",
-                    "STAF",
-                    "MSBF",
-                    "MMV",
-                    "ISM",
-                    "GAF",
-                    "QCCO",
-                    "NEN",
-                    "SPXH",
-                    "NGHCP",
-                    "GNOW",
-                    "SKOR",
-                    "EXT",
-                    "DNBF",
-                    "APP",
-                    "GFED",
-                    "DBU",
-                    "IBDD",
-                    "JMPC",
-                    "IESM",
-                    "SMMU",
-                    "BRID",
-                    "ESEA",
-                    "OBCI",
-                    "AIW",
-                    "EUDG",
-                    "SCIN",
-                    "BPOPM",
-                    "FCE.B",
-                    "IBDB",
-                    "USMD",
-                    "MLPG",
-                    "NFEC",
-                    "FDIV",
-                    "RTR",
-                    "DGT",
-                    "SOP",
-                    "CFA",
-                    "GFNCP",
-                    "JHDG",
-                    "BKJ",
-                    "OSN",
-                    "FCAP",
-                    "MCQ",
-                    "IPD",
-                    "THTI",
-                    "PZI",
-                    "XLRE",
-                    "FLAG",
-                    "GBR",
-                    "ARKK",
-                    "NSEC",
-                    "IPN",
-                    "KEQU",
-                    "VEGA",
-                    "BOM",
-                    "FINZ",
-                    "MCRO",
-                    "ARKG",
-                    "ISDR",
-                    "PXLV",
-                    "NPN",
-                    "GAI",
-                    "SGAR",
-                    "GOODP",
-                    "TCCO",
-                    "UHN",
-                    "ATLC",
-                    "ESCRP",
-                    "DYY",
-                    "INC",
-                    "TCRX",
-                    "FORK",
-                    "TRT",
-                    "CLACU",
-                    "MELR",
-                    "ICBK",
-                    "YCL",
-                    "IKNX",
-                    "FMF",
-                    "HTGZ",
-                    "FCOR",
-                    "UGE",
-                    "JEM",
-                    "WITE",
-                    "MESO",
-                    "TUSA",
-                    "SPYB",
-                    "WBB",
-                    "DOD",
-                    "TAIT",
-                    "BCOM",
-                    "SSBI",
-                    "CUBS",
-                    "JPEM",
-                    "AND",
-                    "SOIL",
-                    "TVIZ",
-                    "VPCOU",
-                    "HBANP",
-                    "JHMM",
-                    "IBMJ",
-                    "QINC",
-                    "NTL",
-                    "WREI",
-                    "LND",
-                    "YPRO",
-                    "ESBK",
-                    "INY",
-                    "WHFBL",
-                    "EQWL",
-                    "QEMM",
-                    "AADR",
-                    "FXS",
-                    "BLVDU",
-                    "GLBS",
-                    "CHOC",
-                    "IIN",
-                    "EVP",
-                    "CIG.C",
-                    "COWNL",
-                    "NTIC",
-                    "MHLDO",
-                    "SBB",
-                    "WMLP",
-                    "SAJA",
-                    "SFBC",
-                    "LGL",
-                    "FCAN",
-                    "BDL",
-                    "RPX",
-                    "ESP",
-                    "MTGEP",
-                    "KAP",
-                    "KBSF",
-                    "ABEOW",
-                    "EMSH",
-                    "CHIE",
-                    "IGIH",
-                    "CFRXW",
-                    "IDHQ",
-                    "MXC",
-                    "ARKQ",
-                    "OEUR",
-                    "LSBK",
-                    "MBFIP",
-                    "KSA",
-                    "GLADO",
-                    "TOFR",
-                    "CNY",
-                    "FUTS",
-                    "TCFC",
-                    "MOC",
-                    "IBCD",
-                    "ARL",
-                    "ONVI",
-                    "SCTO",
-                    "EPRO",
-                    "MIW",
-                    "LBIX",
-                    "EMI",
-                    "CDOR",
-                    "EQWM",
-                    "BHV",
-                    "PSCM",
-                    "JHMF",
-                    "EVBN",
-                    "HRT",
-                    "CHNB",
-                    "WAYN",
-                    "FCCO",
-                    "SDP",
-                    "ASUR",
-                    "GDJS",
-                    "PBM",
-                    "KURA",
-                    "NECB",
-                    "IBMK",
-                    "IBCE",
-                    "DIT",
-                    "TYTE",
-                    "HWBK",
-                    "TCI",
-                    "JHMT",
-                    "JGBS",
-                    "GYC",
-                    "MKC.V",
-                    "DPW",
-                    "NEWTZ",
-                    "PPSI",
-                    "PBSK",
-                    "JHMH",
-                    "RITTW",
-                    "PHII",
-                    "MRGR",
-                    "SIVBO",
-                    "ARY",
-                    "EMCG",
-                    "BRXX",
-                    "OASI",
-                    "FCLF",
-                    "NSAT",
-                    "RCG",
-                    "IPF",
-                    "HCJ",
-                    "BCM",
-                    "XNY",
-                    "TIPX",
-                    "QDEU",
-                    "GLBZ",
-                    "QDYN",
-                    "ASBB",
-                    "CVR",
-                    "FONE",
-                    "IEIS",
-                    "SIZE",
-                    "WYDE",
-                    "BZM",
-                    "CN",
-                    "WMW",
-                    "NSYS",
-                    "GOODO",
-                    "ICVT",
-                    "QVM",
-                    "JSC",
-                    "QBAK",
-                    "MBSD",
-                    "HFEZ",
-                    "ADRU",
-                    "EQFN",
-                    "CNDA",
-                    "TORM",
-                    "SGB",
-                    "CZFC",
-                    "SPLX",
-                    "EQLT",
-                    "HFXJ",
-                    "GWGH",
-                    "CHNA",
-                    "ICN",
-                    "MLTI",
-                    "KBWI",
-                    "ZIONW",
-                    "SGRP",
-                    "MPB",
-                    "FMK",
-                    "XVZ",
-                    "RVNU",
-                    "GGAC",
-                    "FTHI",
-                    "CROP",
-                    "GPIC",
-                    "CPI",
-                    "GDEF",
-                    "MLPS",
-                    "TRSK",
-                    "MES",
-                    "NAIL",
-                    "BPFHP",
-                    "FTT",
-                    "CXSE",
-                    "INR",
-                    "JJN",
-                    "GDJJ",
-                    "TAPR",
-                    "BRN",
-                    "XTNT",
-                    "XKE",
-                    "KTN",
-                    "SDD",
-                    "HAKK",
-                    "TATT",
-                    "AGF",
-                    "CLWT",
-                    "TOTS",
-                    "BOTJ",
-                    "KTEC",
-                    "SDEM",
-                    "TFLO",
-                    "LRGF",
-                    "FSFG",
-                    "IROQ",
-                    "USV",
-                    "FCA",
-                    "PIY",
-                    "FWDB",
-                    "NASH",
-                    "FWDD",
-                    "GOODN",
-                    "CIX",
-                    "APWC",
-                    "FFHL",
-                    "JPN",
-                    "CUBN",
-                    "BUR",
-                    "IRV",
-                    "HHYX",
-                    "VLSM",
-                    "LGLV",
-                    "PAACU",
-                    "GRU",
-                    "GRID",
-                    "WEBK",
-                    "CDL",
-                    "CFNB",
-                    "KTH",
-                    "MATH",
-                    "IBDA",
-                    "SAA",
-                    "PFBX",
-                    "BYLD",
-                    "FORTY",
-                    "AMS",
-                    "SVT",
-                    "GJV",
-                    "GJH",
-                    "DWAT",
-                    "GTAA",
-                    "PSBH",
-                    "FCFP",
-                    "WAFDW",
-                    "VQTS",
-                    "UTES",
-                    "WHLRP",
-                    "LLSP",
-                    "PAF",
-                    "SZK",
-                    "CART",
-                    "EMCF",
-                    "CMDT",
-                    "HSCZ",
-                    "RJZ",
-                    "EQWS",
-                    "GRVY",
-                    "JHMC",
-                    "TSRI",
-                    "VRTB",
-                    "MBCN",
-                    "QAT",
-                    "TONS",
-                    "XTLB",
-                    "HBK",
-                    "TPYP",
-                    "CNR",
-                    "QCAN",
-                    "UNAM",
-                    "WBKC",
-                    "IDXJ",
-                    "IRDMB",
-                    "DIVC",
-                    "EFO",
-                    "BYFC",
-                    "GRF",
-                    "BOXC",
-                    "BICK",
-                    "HYND",
-                    "SMIT",
-                    "PNBK",
-                    "JBR",
-                    "JPUS",
-                    "OLEM",
-                    "UCD",
-                    "HFBL",
-                    "PRGNL",
-                    "PRB",
-                    "ZX",
-                    "ELSE",
-                    "IPK",
-                    "JGBT",
-                    "HLG",
-                    "HAKD",
-                    "IBCC",
-                    "FISK",
-                    "HCACU",
-                    "GTIP",
-                    "IEH",
-                    "JJM",
-                    "MICT",
-                    "CAMBW",
-                    "PNRG",
-                    "AUMAW",
-                    "DELT",
-                    "JCTCF",
-                    "CID",
-                    "TYD",
-                    "EFFE",
-                    "DBAP",
-                    "CFBK",
-                    "IOT",
-                    "WHLM",
-                    "MMTM",
-                    "EQBK",
-                    "CADC",
-                    "EMHZ",
-                    "GCE",
-                    "BANFP",
-                    "VRTA",
-                    "HJV",
-                    "FTW",
-                    "STXX",
-                    "CAMB",
-                    "QJPN",
-                    "HTF",
-                    "BDD",
-                    "GCBC",
-                    "FLN",
-                    "EEML",
-                    "VBFC",
-                    "FDTS",
-                    "WSFSL",
-                    "RWV",
-                    "PJS",
-                    "EVX",
-                    "CUO",
-                    "CAPE",
-                    "DXI",
-                    "INS",
-                    "ALTY",
-                    "KLDW",
-                    "BUNL",
-                    "DIVI",
-                    "CAPX",
-                    "AIII",
-                    "STZ.B",
-                    "RWG",
-                    "DTUS",
-                    "BDMS",
-                    "DPST",
-                    "GJT",
-                    "HMNY",
-                    "FUD",
-                    "FIBG",
-                    "EACQU",
-                    "PYT",
-                    "PAK",
-                    "CPER",
-                    "HCAPL",
-                    "HGJP",
-                    "ROIQ",
-                    "SYE",
-                    "SMK",
-                    "HERO",
-                    "IRCP",
-                    "FTLB",
-                    "LQDH",
-                    "SMCP",
-                    "EEH",
-                    "LTL",
-                    "LKOR",
-                    "SOHOM",
-                    "MEILZ",
-                    "SKLNU",
-                    "FWV",
-                    "RAND",
-                    "SLTC",
-                    "SPXE",
-                    "PILL",
-                    "MOG.B",
-                    "IHT",
-                    "HGEU",
-                    "SPYD",
-                    "DXJC",
-                    "IOIL",
-                    "IQLT",
-                    "SCQ",
-                    "RDIB",
-                    "AUMAU",
-                    "HDAW",
-                    "TFG",
-                    "HECO",
-                    "UBC",
-                    "DAEG",
-                    "FBZ",
-                    "INTF",
-                    "FBSS",
-                    "TIGR",
-                    "URR",
-                    "HOTRW",
-                    "TIL",
-                    "PRSNW",
-                    "CNTR",
-                    "TAGS",
-                    "MOTI",
-                    "AXPWW",
-                    "ROAM",
-                    "KROO",
-                    "MCBK",
-                    "ASHX",
-                    "FMBH",
-                    "OPGNW",
-                    "CAPNW",
-                    "FEUZ",
-                    "ZSML",
-                    "USFR",
-                    "XSOE",
-                    "CCXE",
-                    "OEUH",
-                    "CLAW",
-                    "JJP",
-                    "GVT",
-                    "RWXL",
-                    "CCX",
-                    "UOP",
-                    "SCIU",
-                    "MDVXW",
-                    "TBZ",
-                    "HDWM",
-                    "IJNK",
-                    "AFM",
-                    "GTN.A",
-                    "VIIZ",
-                    "LOWC",
-                    "AXX",
-                    "IBCB",
-                    "DTRM",
-                    "ICSH",
-                    "NXTDW",
-                    "GGACR",
-                    "ACTX",
-                    "FFBCW",
-                    "BIO.B",
-                    "KEN",
-                    "ACBI",
-                    "HAHA",
-                    "ROUS",
-                    "SIBC",
-                    "LDRI",
-                    "SVLC",
-                    "PBCP",
-                    "WHLRW",
-                    "MAYS",
-                    "LSC",
-                    "CVB",
-                    "MDSY",
-                    "CSB",
-                    "CMD",
-                    "JPX",
-                    "PGM",
-                    "RIBTW",
-                    "GJS",
-                    "QMEX",
-                    "WSO.B",
-                    "MATL",
-                    "TYBS",
-                    "SBUS",
-                    "IDHB",
-                    "ISCF",
-                    "EMMSP",
-                    "HHDG",
-                    "ATRM",
-                    "WYIGU",
-                    "SMLF",
-                    "EMRE",
-                    "AAPC",
-                    "HDLS",
-                    "LMFAU",
-                    "TAP.A",
-                    "LBTYB",
-                    "BWINA",
-                    "GIVE",
-                    "SOHOL",
-                    "CKX",
-                    "BDAT",
-                    "NUROW",
-                    "GTWN",
-                    "USATP",
-                    "GRSHU",
-                    "HREX",
-                    "HDEF",
-                    "CSMA",
-                    "HYIH",
-                    "HKOR",
-                    "AXJV",
-                    "IPOS",
-                    "FULLL",
-                    "WVFC",
-                    "IMTM",
-                    "CSA",
-                    "BITE",
-                    "WDRW",
-                    "CIL",
-                    "ACWF",
-                    "UMX",
-                    "UCIB",
-                    "JJS",
-                    "SIPE",
-                    "SCC",
-                    "UBN",
-                    "HDRAW",
-                    "SMAC",
-                    "JJU",
-                    "MLPV",
-                    "SBOT",
-                    "CADTR",
-                    "RILY",
-                    "GGACW",
-                    "HEWU",
-                    "VIXH",
-                    "WINS",
-                    "CDORP",
-                    "UAG",
-                    "BGI",
-                    "MRRL",
-                    "XGTIW",
-                    "DXJT",
-                    "IPB",
-                    "QTNTW",
-                    "DISCB",
-                    "ECACU",
-                    "HDWX",
-                    "HCACW",
-                    "UBIC",
-                    "EMAG",
-                    "COBO",
-                    "SPMO",
-                    "GJR",
-                    "JDG",
-                    "THHY",
-                    "QWLD",
-                    "GALTW",
-                    "CLYH",
-                    "EWRE",
-                    "AFCB",
-                    "ALLB",
-                    "WEET",
-                    "GPACU",
-                    "EMCR",
-                    "QVCB",
-                    "KCC",
-                    "GGACU",
-                    "DXUS",
-                    "STPP",
-                    "SBM",
-                    "KLEM",
-                    "BSCP",
-                    "SAGG",
-                    "JJE",
-                    "LSTK",
-                    "HDRA",
-                    "JW.B",
-                    "GJO",
-                    "FTSD",
-                    "QKOR",
-                    "GMMB",
-                    "EMFT",
-                    "RNVA",
-                    "HBU",
-                    "RODM",
-                    "FAUS",
-                    "ITEQ",
-                    "BDG",
-                    "BWV",
-                    "DHXM",
-                    "CHEP",
-                    "EZR",
-                    "XINA",
-                    "ELECU",
-                    "QEFA",
-                    "EPIX",
-                    "DVHI",
-                    "ZTEST",
-                    "USSD",
-                    "FMLP",
-                    "MULT",
-                    "MGYR",
-                    "HUSE",
-                    "CSBR",
-                    "CLRBW",
-                    "AGGY",
-                    "HEGE",
-                    "JXSB",
-                    "ERO",
-                    "CDORO",
-                    "IGS",
-                    "TFSCW",
-                    "SBV",
-                    "QLC",
-                    "DDP",
-                    "DIRT",
-                    "BCHP",
-                    "TROVW",
-                    "XCOM",
-                    "BOS",
-                    "CINR",
-                    "BRAF",
-                    "HEWL",
-                    "BVXVW",
-                    "SENEB",
-                    "EMIH",
-                    "SIZ",
-                    "MDGS",
-                    "CHEKW",
-                    "HHFR",
-                    "HRMNW",
-                    "HBZ",
-                    "AMIC",
-                    "HDRW",
-                    "CIVBP",
-                    "MEILW",
-                    "PXS",
-                    "ARWA",
-                    "ELECW",
-                    "EMSD",
-                    "EUDV",
-                    "UJB",
-                    "BPFHW",
-                    "DBUK",
-                    "STXSW",
-                    "QGTA",
-                    "KEMP",
-                    "GPIA",
-                    "XVIX",
-                    "CSCB",
-                    "HDRAU",
-                    "FCFI",
-                    "HAWX",
-                    "CADTW",
-                    "ROGS",
-                    "SPUN",
-                    "OAPH",
-                    "JUNR",
-                    "OXBRW",
-                    "CRDT",
-                    "WIL",
-                    "ELLO",
-                    "STRZB",
-                    "AZIA",
-                    "XCEM",
-                    "FXSG",
-                    "SPXV",
-                    "QPAC",
-                    "LBND",
-                    "GHII",
-                    "ZVV",
-                    "CUPM",
-                    "QESP",
-                    "HEUS",
-                    "TRTL",
-                    "BDCZ",
-                    "UBG",
-                    "GURI",
-                    "DEE",
-                    "PILS",
-                    "HEMV",
-                    "FGBI",
-                    "DPU",
-                    "ZAZA",
-                    "HEUV",
-                    "KFYP",
-                    "DRVN",
-                    "QPACU",
-                    "AUMA",
-                    "FLAT",
-                    "GPIAU",
-                    "QUS",
-                    "FXCH",
-                    "DBIF",
-                    "SPGH",
-                    "DLBL",
-                    "PAACR",
-                    "ICI",
-                    "FUE",
-                    "DELTW",
-                    "SCIX",
-                    "SAUC",
-                    "HACV",
-                    "HEFV",
-                    "VLU",
-                    "FIA",
-                    "SBEU",
-                    "SYV",
-                    "TRTLU",
-                    "DTYL",
-                    "BSJN",
-                    "CTNN",
-                    "CFGE",
-                    "UBND",
-                    "SLOW",
-                    "ICLDW",
-                    "CNLMW",
-                    "AGM.A",
-                    "DBMX",
-                    "PFEM",
-                    "FKO",
-                    "SMLL",
-                    "SMACR",
-                    "RORO",
-                    "GWRU",
-                    "GPIAW",
-                    "VLLV",
-                    "FEMB",
-                    "GJP",
-                    "MAUI",
-                    "MDLL",
-                    "OFED",
-                    "LBDC",
-                    "CSCR",
-                    "ELEC",
-                    "TFSCU",
-                    "GPACW",
-                    "DIVA",
-                    "LTRPB",
-                    "QTWN",
-                    "MICTW",
-                    "BHAC",
-                    "LVNTB",
-                    "DFVS",
-                    "BNTCW",
-                    "TRTLW",
-                    "DFVL",
-                    "VLML",
-                    "UXJ",
-                    "PAACW",
-                    "EEHB",
-                    "HEGJ",
-                    "EAGL",
-                    "QEH",
-                    "HVT.A",
-                    "CNLM",
-                    "SPXT",
-                    "TYNS",
-                    "QPACW",
-                    "DTUL",
-                    "PUTX",
-                    "GMTB",
-                    "ADZ",
-                    "HOLD",
-                    "FOIL",
-                    "ZTST",
-                    "RISE",
-                    "IGU",
-                    "HMG",
-                    "SGYPU",
-                    "MTBCP",
-                    "HEWW",
-                    "HOMX",
-                    "FNTCU",
-                    "ONG",
-                    "GSIE",
-                    "GURX",
-                    "HDRAR",
-                    "JBK",
-                    "ARWAW",
-                    "KRS",
-                    "OBAS",
-                    "HJPX",
-                    "SCID",
-                    "COYN",
-                    "EMBH",
-                    "HEWY",
-                    "AGA",
-                    "CCZ",
-                    "GALTU",
-                    "ZLRG",
-                    "UBM",
-                    "PGD",
-                    "JJT",
-                    "ZIONZ",
-                    "CERC",
-                    "FCVT",
-                    "ALFI",
-                    "HRMNU",
-                    "USLB",
-                    "ARWAU",
-                    "DCNG",
-                    "UK",
-                    "BHACU",
-                    "AMUB",
-                    "CAMBU",
-                    "CNLMU",
-                    "CNHX",
-                    "JPNH",
-                    "OSMS",
-                    "FLRT",
-                    "RGRO",
-                    "SLDR",
-                    "USAG",
-                    "CBO",
-                    "JHA",
-                    "LMCB",
-                    "HEWC",
-                    "HCAC",
-                    "SBFGP",
-                    "TFSC",
-                    "SPXN",
-                    "EMBB",
-                    "QGBR",
-                    "NINI",
-                    "JGBB",
-                    "CADTU",
-                    "MLPO",
-                    "LEDD",
-                    "GBB",
-                    "STRI",
-                    "PACE",
-                    "GPAC",
-                    "AFQ",
-                    "GSD",
-                    "PRME",
-                    "SYTL",
-                    "TFSCR",
-                    "COYNW",
-                    "DGICB",
-                    "CYCCP",
-                    "BHACW",
-                    "IDLB",
-                    "USWD",
-                    "HAUD",
-                    "IAGG",
-                    "TLDH",
-                    "TLEH",
-                    "KELYB",
-                    "MLPB",
-                    "PAAC",
-                    "SCIJ",
-                    "CFCB",
-                    "LLEM",
-                    "ARWAR",
-                    "LLSC",
-                    "FNTC",
-                    "LD",
-                    "EAGLW",
-                    "IKAN",
-                    "CATYW",
-                    "DBIZ",
-                    "RPRXZ",
-                    "JYN",
-                    "HDEZ",
-                    "FXEP",
-                    "HRMN",
-                    "SBNYW",
-                    "CSWIV",
-                    "WVVIP",
-                    "CNLMR",
-                    "LLDM",
-                    "WTFCW",
-                    "CADT",
-                    "IVLU",
-                    "SPVU",
-                    "DBAU",
-                    "GRWN",
-                    "HACW",
-                    "AYT",
-                    "QED",
-                    "DBRE",
-                    "QLS",
-                    "XXV",
-                    "BHACR",
-                    "BLNG",
-                    "BLVD",
-                    "BLVDW",
-                    "CERCW",
-                    "CERCZ",
-                    "CSWCV",
-                    "DBSE",
-                    "EACQ",
-                    "EACQW",
-                    "ECAC",
-                    "ECACR",
-                    "EMSA",
-                    "ERW",
-                    "FNTCW",
-                    "FXJP",
-                    "GRSH",
-                    "GRSHW",
-                    "HEVY",
-                    "ISZE",
-                    "JGBL",
-                    "PACEW",
-                    "PXSSV",
-                    "RODI",
-                    "SMACU",
-                    "TZF",
-                    "WYIG",
-                    "WYIGW",
-                };
-
-                #endregion
-            }
-        }
-    }
-}
