+++ Algorithm.CSharp/Benchmarks/EmptyMinute400EquityBenchmark.cs
@@ -1,11 +1,48 @@
-﻿using System;
+﻿/*
+ * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
+ * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using QuantConnect.Data;
 
 namespace QuantConnect.Algorithm.CSharp.Benchmarks
 {
+    /// <summary>
+    /// Benchmark Algorithm: Loading and synchronization of 400 equity minute symbols.
+    /// </summary>
+    /// <remarks>
+    /// Testing the parsing; synchronizing and injection of data into LEAN event handlers. This directly adds the symbols to avoid universe
+    /// selection overhead. Later tests will include universe selection. The data sources are sparse.
+    /// </remarks>
+    public class EmptyMinute400EquityBenchmark : QCAlgorithm
+    {
+        public override void Initialize()
+        {
+            SetStartDate(2015, 10, 1);
+            SetEndDate(2015, 11, 15);
+            foreach (var symbol in Symbols.Equity.All.Take(400))
+            {
+                AddSecurity(SecurityType.Equity, symbol);
+            }
+        }
+
+        public override void OnData(Slice slice)
+        {
+        }
+    }
+
     public static class Symbols
     {
         public static class Equity
